### 2025-09-05 #8934 @thosvarley
This work was done while working on a Department of Defense contract that wouldn't let me publish anything publicly until I was approved to do so. I just got approval earlier this week and moved everything _en masse_ to a new Github repo. I've been developing this package for the last two years.
