### 2025-09-05 d805471
Fixing paper YAML
