+++ joss_paper/paper.md
@@ -28,7 +28,7 @@ The Syntropy package is broken up into two main arms: discrete estimators (which
 Wherever possible, we have tried to ensure overlap between the two branches, ensuring that each information theoretic function has both discrete and continuous implementations (e.g. there is a discrete and continuous entropy estimator, a discrete and continuous mutual information estimator, etc). This ensures that researchers can select the most appropriate tool for the data they have, rather than being forced to transform their data to make it fit (e.g. discretizing continuous data). 
 
 # Statement of need
-While many different computing packages for information-theoretic analysis exist, they are typically restricted to only a single subset of analyses or data type. For example, the DIT package [@james_dit_2018] is specific to discrete information theory and uses a customized distribution object. The IDTxl [@wollstadt_idtxl_2019] package implements several different classes of estimators, but is focused specifically on information dynamics and time series analysis, as is the JIDT package [@lizier_jidt_2014]. Information decomposition analysis in particular is an outstanding issue: DIT and IDTxl have limited support for the partial information decomposition (PID) of discrete random variables, but no other package also implements the PID for continuous Gaussian random variables, or supports the partial entropy decomposition (PED), generalied information decomposition (GID), or alpha-synergy decomposition. Several other classes of analyses are also included in Syntropy that are not present in other primary packages, including spectral estimators for Gaussian autoregressive processes, and Lempel-Ziv estimators for discrete dynamic processes. 
+While many different computing packages for information-theoretic analysis exist, they are typically restricted to only a single subset of analyses or data type. For example, the DIT package [@james_dit_2018] is specific to discrete information theory and uses a customized distribution object. The IDTxl [@wollstadt_idtxl_2019] package implements several different classes of estimators, but is focused specifically on information dynamics and time series analysis, as is the JIDT package [@lizier_jidt_2014]. Information decomposition analysis in particular is an outstanding issue: DIT and IDTxl have limited support for the partial information decomposition (PID) of discrete random variables, but no other package also implements the PID for continuous Gaussian random variables, or supports the partial entropy decomposition (PED), generalized information decomposition (GID), or alpha-synergy decomposition. Several other classes of analyses are also included in Syntropy that are not present in other primary packages, including spectral estimators for Gaussian autoregressive processes, and Lempel-Ziv estimators for discrete dynamic processes. 
  
 # Acknowledgments 
 This work was supported by the Cold Regions Research and Engineering Laboratory, under contract number W913E524C0012. TFV would also like to thank Dr. Olaf Sporns and Dr. Maria Pope for extensive support and discussion about all things information theory over many years, and Dr. Josh Bongard for support while finalizing this package. 
