+++ joss_paper/paper.md
@@ -11,7 +11,7 @@ authors:
 affiliations:
     - name: Vermont Complex Systems Institute, University of Vermont, Burlington, VT, USA. 
     index: 1
-date: 1 May 2025
+date: 01 May 2025
 bibliography: references.bib
 ---
 
