### 2025-09-06 #8934 @crvernon
@thosvarley - Thanks for the clarification.  Your header format is still incorrect.  Note the missing `-` in front of `index` in your affiliations section of the header.
