### 2025-09-10 918585e
Updating JOSS paper + allow singular matrices
