### 2025-09-17 da0a6a0
Updating paper.md
