### 2025-09-17 6507156
Updating paper.md
