+++ joss_paper/paper.md
@@ -6,11 +6,11 @@ tags:
     - complex systems 
 authors:
     - name: Thomas F. Varley
-      orcid: 0000-0002-3317-9882 
-      affiliation: "1"
+    - orcid: 0000-0002-3317-9882 
+    - affiliation: 1
 affiliations:
     - name: Vermont Complex Systems Institute, University of Vermont, Burlington, VT, USA. 
-      index: 1
+    - index: 1
 date: 01 May 2025
 bibliography: references.bib
 ---
