### 2025-09-17 b3ecd9f
Debugging paper
