+++ joss_paper/paper.md
@@ -16,19 +16,37 @@ bibliography: references.bib
 ---
 
 # Summary:
-Information theory (a branch of mathematics concerned with the logic of inference and uncertainty) has emerged as a kind of *lingua franca* for the study of complex systems [@varley_information_2024], finding applications in neuroscience [@timme_synergy_2014,@varley_serotonergic_2024], climatology [@goodwell_its_2020,@goodwell_debatesdoes_2020], developmental biology [@blackiston_revealing_2025], economics [@kim_predicting_2020,@rajpal_synergistic_2025], and more. The goal of the Syntropy package is to provide an accessible, easy-to-use API for scientists interested in applying information-theoretic analyses to arbitrary data sets. To this end, Syntropy implements a variety of different measures, as well as a variety of estimators for different types of data (discrete, continuous, time series, etc). Being primarily concerned with multivariate information theory and the application to complex systems, a core part of the Syntropy package is implementing the various information decompositions and higher-order mutual information measures that have been developed in recent years for the study of redundant and synergistic interactions in complex systems. 
+Information theory (a branch of mathematics concerned with the logic of inference and uncertainty) has emerged as a kind of *lingua franca* for the study of complex systems [@varley_information_2024], finding applications in neuroscience [@timme_synergy_2014][@varley_serotonergic_2024], climatology [@goodwell_its_2020][@goodwell_debatesdoes_2020], developmental biology [@blackiston_revealing_2025], economics [@kim_predicting_2020][@rajpal_synergistic_2025], sociology [varley_untangling_2022] and more. 
+The goal of the Syntropy package is to provide an accessible, easy-to-use API for scientists interested in applying information-theoretic analyses to arbitrary data sets. 
+To this end, Syntropy implements a variety of different measures, as well as a variety of estimators for different types of data (discrete, continuous, time series, etc). 
+Being primarily concerned with multivariate information theory and the application to complex systems, a core part of the Syntropy package is implementing the various information decompositions and higher-order mutual information measures that have been developed in recent years for the study of redundant and synergistic interactions in complex systems. 
 
-The Syntropy package takes considerable inspiration from the Networkx package [@hagberg_exploring_2008]. By providing an accessible, high-level interface for complex analyses, the developers of Networkx have made network science and graph-theoretic tools available to a much more diverse range of scientists than would ever have been possible if every researchers was required to develop their own functions. The goal of Syntropy is to be the Networkx of information theory"; lowering the barrier to entry fro information theory so that powerful tools are accessible to scientists from many different fields. Accordingly, the Syntropy package is written in pure Python, with minimal external dependencies, all of which are standard libraries for scientific computing in Python, including ``numpy``, ``scipy``, and ``networkx`` packages. 
+The Syntropy package takes considerable inspiration from the Networkx package [@hagberg_exploring_2008]. 
+By providing an accessible, high-level interface for complex analyses, the developers of Networkx have made network science and graph-theoretic tools available to a much more diverse range of scientists than would ever have been possible if every researchers was required to develop their own functions. 
+The goal of Syntropy is to be the Networkx of information theory"; lowering the barrier to entry fro information theory so that powerful tools are accessible to scientists from many different fields. 
+Accordingly, the Syntropy package is written in pure Python, with minimal external dependencies, all of which are standard libraries for scientific computing in Python, including ``numpy``, ``scipy``, and ``networkx`` packages. 
 
 While the core information-theoretic functions have been implemented and tested for correctness, the goal is for Syntropy to be a "living package": updating the library of functions and estimators as the field of multivariate information theory continues to develop. Future work may also include porting the library to other scientific programming languages such as Julia, R, or MATLAB. 
 
 # Description of package
[…]
