+++ joss_paper/references.bib
@@ -1,3 +1,25 @@
+@article{varley_untangling_2022,
+	title = {Untangling {Synergistic} {Effects} of {Intersecting} {Social} {Identities} with {Partial} {Information} {Decomposition}},
+	volume = {24},
+	copyright = {http://creativecommons.org/licenses/by/3.0/},
+	issn = {1099-4300},
+	url = {https://www.mdpi.com/1099-4300/24/10/1387},
+	doi = {10.3390/e24101387},
+	abstract = {The theory of intersectionality proposes that an individual’s experience of society has aspects that are irreducible to the sum of one’s various identities considered individually, but are “greater than the sum of their parts”. In recent years, this framework has become a frequent topic of discussion both in social sciences and among popular movements for social justice. In this work, we show that the effects of intersectional identities can be statistically observed in empirical data using information theory, particularly the partial information decomposition framework. We show that, when considering the predictive relationship between various identity categories such as race and sex, on outcomes such as income, health and wellness, robust statistical synergies appear. These synergies show that there are joint-effects of identities on outcomes that are irreducible to any identity considered individually and only appear when specific categories are considered together (for example, there is a large, synergistic effect of race and sex considered jointly on income irreducible to either race or sex). Furthermore, these synergies are robust over time, remaining largely constant year-to-year. We then show using synthetic data that the most widely used method of assessing intersectionalities in data (linear regression with multiplicative interaction coefficients) fails to disambiguate between truly synergistic, greater-than-the-sum-of-their-parts interactions, and redundant interactions. We explore the significance of these two distinct types of interactions in the context of making inferences about intersectional relationships in data and the importance of being able to reliably differentiate the two. Finally, we conclude that information theory, as a model-free framework sensitive to nonlinearities and synergies in data, is a natural method by which to explore the space of higher-order social dynamics.},
+	language = {en},
+	number = {10},
+	urldate = {2022-09-28},
+	journal = {Entropy},
+	author = {Varley, Thomas F. and Kaminski, Patrick},
+	month = oct,
+	year = {2022},
+	note = {Number: 10
+Publisher: Multidisciplinary Digital Publishing Institute},
+	keywords = {partial information decomposition, synergy, intersectionality, higher-order interactions, sociology},
+	pages = {1387},
+	annote = {Number: 10 Publisher: Multidisciplinary Digital Publishing Institute},
+	file = {Full Text PDF:/home/thosvarley/Zotero/storage/3SDTWYF8/Varley and Kaminski - 2022 - Untangling Synergistic Effects of Intersecting Soc.pdf:application/pdf;Snapshot:/home/thosvarley/Zotero/storage/5SVX8NM3/1387.html:text/html;Snapshot:/home/thosvarley/Zotero/storage/2YRMCB3N/1387.html:text/html},
+}
 
 @article{timme_synergy_2014,
 	title = {Synergy, redundancy, and multivariate information measures: an experimentalist’s perspective},
