### 2025-09-17 4f071d4
Debugging paper
