### 2025-09-17 8ff15ce
Update paper.md
