### 2025-10-12 1e81ae4
Improving cov indexing
