+++ syntropy/gaussian/temporal.py
@@ -5,6 +5,50 @@ import scipy.io as io
 import scipy.stats as stats
 
 
+def construct_csd_tensor(
+    idxs: tuple, data: np.ndarray, fs: int = 1, nperseg: int = 1024
+) -> np.ndarray:
+    """
+
+    Parameters
+    ----------
+    idxs : tuple
+
+    data : np.ndarray
+
+    fs : int
+
+    nperseg : int
+
+
+    Returns
+    -------
+    np.ndarray
+
+
+    """
+    N: int = len(idxs)
+    S: np.ndarray = np.zeros((nperseg, N, N))
+
+    for i in range(N):
+        for j in range(i + 1):
+            f, Pij = signal.csd(
+                data[idxs[i], :],
+                data[idxs[j], :],
+                return_onesided=False,
+                fs=fs,
+                nperseg=nperseg,
+            )
+            Pij = np.fft.fftshift(Pij)
+            S[:i, j] = Pij.real
+            S[:, j, i] = np.conj(Pij).real
+
+    omega = 2 * np.pi * f / fs
+    omega = np.fft.fftshift(omega)
+
+    return S, omega
+
+
 def differential_entropy_rate(
     idxs: tuple,
     data: np.ndarray,
@@ -40,23 +84,7 @@ def differential_entropy_rate(
     # %%
     N = len(idxs)
 
-    S = np.zeros((nperseg, N, N))
-
-    for i in range(N):
-        for j in range(i + 1):
-            f, Pij = signal.csd(
-                data[idxs[i], :],
-                data[idxs[j], :],
-                return_onesided=False,
-                fs=fs,
-                nperseg=nperseg,
-            )
-            Pij = np.fft.fftshift(Pij)
-            S[:, i, j] = Pij.real
-            S[:, j, i] = np.conj(Pij).real
-
-    omega = 2 * np.pi * f / fs  # from Hz to rad/sample
-    omega = np.fft.fftshift(omega)
+    S, omega = construct_csd_tensor(idxs=idxs, data=data, fs=fs, nperseg=nperseg)
 
     ptw = np.array(
         [(1 / 2) * np.log(((2 * np.pi * np.e) ** N) * np.linalg.det(a)) for a in S]
@@ -72,10 +100,10 @@ def mutual_information_rate(
     data: np.ndarray,
     fs: int = 1,
     nperseg: int = 1024,
-) -> (np.ndarray, float):
+) -> (np.ndarray, float, np.ndarray):
     """
     Computes the mutual information rate between two (potentially multivariate) Gaussian processes.
-    
+
     :math:`I(X; Y) = \\frac{1}{4\\pi} \\int_{-\\pi}^{\\pi} \\log \\left( \\frac{  |S_X(\\omega)||S_Y(\\omega)| }{ |S_{XY}(\\omega)| } \\right) d\\omega`
 
     See:
@@ -121,27 +149,10 @@ def mutual_information_rate(
     reidxs_x: tuple = tuple(i for i in range(Nx))
     reidxs_y: tuple = tuple(i + Nx for i in range(Ny))
 
-    S: np.ndarray = np.zeros((nperseg, N, N))
+    S, omega = construct_csd_tensor(idxs=idxs_, data=data, fs=fs, nperseg=nperseg)
 
-    for i in range(N):
-        for j in range(i + 1):
-            f, Pij = signal.csd(
-                data[idxs_[i], :],
-                data[idxs_[j], :],
-                return_onesided=False,
-                fs=fs,
-                nperseg=nperseg,
-            )
-
-            Pij: np.ndarray = np.fft.fftshift(Pij)
-            S[:, i, j] = Pij.real
-            S[:, j, i] = np.conj(Pij).real
-
[…]
