### 2025-10-12 900f3f4
Improving cov indexing
