### 2025-10-12 8fe302a
Normalizing flow-based neural estimator
