+++ syntropy/neural/utils.py
@@ -1,0 +1,197 @@
+import torch
+import torch.optim as optim
+from torch.utils.data import DataLoader, TensorDataset
+import numpy as np
+
+from nflows.distributions.normal import StandardNormal
+from nflows.flows.base import Flow
+from nflows.transforms.autoregressive import MaskedAffineAutoregressiveTransform
+from nflows.transforms.base import CompositeTransform
+from nflows.transforms.permutations import RandomPermutation
+
+
+def initialize_flow(
+    dim: int,
+    dim_context: int = 0,
+    num_layers: int = 5,
+    hidden_features: int = 64,
+    dropout_probability=0.1,
+):
+    """
+    Initializes a new normalizing flow network.
+
+    Parameters
+    ----------
+    dim : int
+        The number of input dimensions.
+    dim_context : int
+        The number of conditioning dimension.
+        The default is zero.
+    num_layers : int
+        The number of hidden layers.
+        The default is 5 layers.
+    hidden_features : int
+        The number of neurons in each hidden layer.
+        The default is 64 neurons.
+    dropout_probability : int
+        The probability of a neuron dropping out.
+
+    Returns
+    -------
+        nflows.flows.base.Flow
+    """
+    transforms: list = []
+
+    for _ in range(num_layers):
+        transforms.append(
+            MaskedAffineAutoregressiveTransform(
+                features=dim,
+                hidden_features=hidden_features,
+                context_features=dim_context,
+                dropout_probability=dropout_probability,
+            )
+        )
+        transforms.append(RandomPermutation(features=dim))
+
+    transform = CompositeTransform(transforms)
+    base_dist = StandardNormal(shape=[dim])
+
+    return Flow(transform, base_dist)
+
+
+def train_flow(
+    flow,
+    data: torch.Tensor,
+    context: None | torch.Tensor = None,
+    batch_size: int = 256,
+    lr: float = 1e-4,
+    num_epochs: int = 100,
+    weight_decay: float = 1e-5,
+    convergence_threshold: float = 0.0,
+    alpha: float = 0.1,
+    verbose: bool = False,
+):
+    """
+    Trains a normalizing flow network to approximate the maximally likely distribution to have generated the given data.
+
+    Parameters
+    ----------
+    flow : nflows.flows.base.Flow
+        An untrained normalizing flow network.
+    data : torch.Tensor
+        The training data, in samples x features format.
+    context : None | torch.Tensor
+        Conditioning random variables.
+        Default is None.
+    batch_size : int
+        The size of each batch.
+        The default value is 256.
+    lr : float
+        The learning rate.
+        The default value is 1e-4
+    num_epochs : int
+        The number of training epochs.
+        The default value is 100,
+    weight_decay: float
+        The rate at which parameter weights decay. A regularizer to reduce over-fitting.
+        The default is 1e-5.
+    convergence_threshold : float
+        The value of the coefficient of variation below which the training terminates.
+        The default value is 0.0, which means the training will not stop before num_epochs is hit.
+    alpha : float
+        How quickly the exponentially weighted moving standard deviation downweights older data.
+        The default is 0.2.
+    verbose : bool
+        Whether to print the loss for each epoch throughout training.
+        The default is False.
+
+
+    Returns
+    -------
+
+    nflows.flows.base.Flow
+    """
+
+    if context is None:
+        dataset = TensorDataset(data)
+    elif context is not None:
+        dataset = TensorDataset(data, context)
+
+    loader = DataLoader(dataset, batch_size=batch_size, shuffle=True)
+
[…]
