### 2025-10-12 7617c66
Normalizing flow-based neural estimator
