### 2025-10-12 75204af
Cleaning neural estimators
