+++ syntropy/neural/shannon.py
@@ -117,47 +117,3 @@ def mutual_information(
     return mi 
 
 
-"""
-in_dir = "/home/thosvarley/Documents/UVM/research/normalizing_flow/data/"
-from syntropy.gaussian.shannon import conditional_mutual_information as cMI
-
-cov: np.ndarray = np.load(in_dir + "six_cov.npz")["arr_0"]
-
-gaussian = stats.multivariate_normal(mean=np.zeros(cov.shape[0]), cov=cov)
-data_train = torch.Tensor(gaussian.rvs(20_000))
-data_test = torch.Tensor(gaussian.rvs(20_000))
-
-
-idxs_x = (0, 1)
-idxs_y = (2, 3)
-idxs_z = (5, 4)
-
-flow_kwargs = {"num_layers": 10, "hidden_features": 64}
-
-train_kwargs = {"num_epochs": 50, "lr": 1e-4,
-                "convergence_threshold": 0.01}
-
-ests = []
-for i in range(10):
-    mi = mutual_information(
-        idxs_x=idxs_x,
-        idxs_y=idxs_y,
-        data=data_train,
-        data_test=data_test,
-        context=idxs_z,
-        verbose=True,
-        flow_kwargs=flow_kwargs,
-        train_kwargs=train_kwargs,
-    )
-    ests.append(mi)
-    print(i)
-
-#h, _ = differential_entropy(
-#    idxs=idxs_x, data=data_train, data_test=data_test, context=idxs_y + idxs_z, verbose=True
-#)
-
-
-
-# print(f"MI: {cMI(idxs_x, idxs_y, idxs_z, cov)} nat.")
-# print(f"Estimated MI: {mi} +/- {stderr}")
-"""
