### 2025-10-14 a3a5b5f
Mypy static typing
