+++ syntropy/discrete/alpha_synergy.py
@@ -3,15 +3,21 @@ import numpy as np
 
 from syntropy.discrete.shannon import conditional_entropy
 from syntropy.discrete.optimization import constrained_maximum_entropy_distributions
-from syntropy.discrete.utils import get_marginal_distribution
+from syntropy.discrete.utils import get_marginal_distribution, reduce_state
+
+from typing import Any
+
+DiscreteDist = dict[tuple[Any, ...], float]
+AlphaSynDist = dict[tuple[Any, ...], float]
+PartialSpectra = dict[tuple[Any, ...], list[float]]
 
 
 def alpha_synergistic_entropy(
-    joint_distribution: dict[tuple, float],
+    joint_distribution: DiscreteDist,
     alpha: int,
     num_samples: int = -1,
     definition: str = "min",
-) -> dict[tuple, float]:
+) -> AlphaSynDist:
     """
     Computes the :math:`\\alpha`-synergistic entropy for a joint distribution
     for a given value of :math:`\\alpha`. See:
@@ -40,19 +46,24 @@ def alpha_synergistic_entropy(
         The local alpha-synergy for each state.
 
     """
+    assert definition in {
+        "min",
+        "max",
+        "avg",
+    }, "The optional definitions are 'min', 'max', or 'avg'."
 
-    N: int = len(list(joint_distribution.keys())[0])
+    N: int = len(next(iter(joint_distribution)))
+
+    val: float = 0.0
 
     if definition == "min":
-        val: float = np.inf
+        val = np.inf
     elif definition == "max":
-        val: float = -np.inf
-    elif val == "average":
-        val: float = 0.0
+        val = -np.inf
 
-    alpha_syns: dict[tuple, float] = {key: val for key in joint_distribution.keys()}
+    alpha_syns: AlphaSynDist = {key: val for key in joint_distribution.keys()}
 
-    sources: list[tuple] = []
+    sources: list[tuple[Any, ...]] = []
     if num_samples == -1:  # Get all combinations of elements of size alpha.
         sources += list(itertools.combinations(tuple(i for i in range(N)), r=alpha))
     else:  # Randomly sample elements of size alpha.
@@ -65,29 +76,31 @@ def alpha_synergistic_entropy(
 
     num_sources: int = len(sources)
 
-    residuals: list = [
+    residuals: list[
+        tuple[Any, ...]
+    ] = [  # The residual indices for each source in sources.
         tuple(i for i in range(N) if i not in source) for source in sources
     ]
 
     for i in range(num_sources):
+        source: tuple[Any, ...] = sources[i]  # The indices of the source variables.
+        complement: tuple[Any, ...] = residuals[
+            i
+        ]  # The indices of the complenetary variables.
 
-        source: tuple = sources[i]
-        complement: tuple = residuals[i]
-
-        idxs: tuple = source + complement
+        idxs: tuple[Any, ...] = source + complement
 
         ptw, _ = conditional_entropy(source, complement, joint_distribution)
 
         for key in ptw.keys():
-
             # Undoing the unpacking of state into source and
             # conditional
-            flatten: tuple = sum(key, ())  # ((0,1),(2,)) -> (0,1,2)
+            flatten: tuple[Any, ...] = sum(key, ())  # ((0,1),(2,)) -> (0,1,2)
 
-            state = [0] * N
-            for i in range(N):
-                state[idxs[i]] = flatten[i]
-            state = tuple(state)
+            temp: list[int] = [0] * N
+            for i in range(N):  # For each element of flatten:
+                temp[idxs[i]] = flatten[i]  # Mapping back to the original order.
+            state: tuple[Any, ...] = tuple(temp)
 
             # Accounting for the various ways synergy can be defined.
[…]
