### 2025-10-18 143aa22
README + type-hints
