### 2025-10-18 2819413
README + type-hints
