+++ syntropy/discrete/shannon.py
@@ -1,10 +1,14 @@
 import numpy as np
 from syntropy.discrete.utils import get_marginal_distribution
 
-def shannon_entropy(joint_distribution: dict) -> tuple[dict, float]:
+from typing import Any
+
+DiscreteDist = dict[tuple[Any, ...], float]
+
+def shannon_entropy(joint_distribution: DiscreteDist) -> tuple[dict, float]:
     """
     Computes the Shannon entropy of the distribution :math:`P(x)`.
-    
+
     .. math::
         H(X) = -\\sum_{x} P(x) \\log P(x)
 
@@ -39,7 +43,7 @@ def shannon_entropy(joint_distribution: dict) -> tuple[dict, float]:
 
 
 def conditional_entropy(
-    inputs_x: tuple, inputs_y: tuple, joint_distribution: dict
+    idxs_x: tuple[int, ...], idxs_y: tuple[int, ...], joint_distribution: DiscreteDist
 ) -> tuple[dict, float]:
     """
     Computes the conditional entropy of X given Y.
@@ -48,9 +52,9 @@ def conditional_entropy(
 
     Parameters
     ----------
-    inputs_x : tuple
+    idxs_x : tuple
         The indices of the variables to compute the entropy on.
-    inputs_y : tuple
+    idxs_y : tuple
         The indicies of the variables to contintue on.
     joint_distribution : dict
         DESCRIPTION.joint_distribution : dict
@@ -64,21 +68,21 @@ def conditional_entropy(
         The pointwise entropy for each state in the joint distribution.
     avg : float
         The average entropy
-    
+
     """
 
-    Nx: int = len(inputs_x)
+    Nx: int = len(idxs_x)
 
-    inputs_xy: tuple = inputs_x + inputs_y
+    idxs_xy: tuple[int, ...] = idxs_x + idxs_y
 
-    marginal_xy: dict = get_marginal_distribution(inputs_xy, joint_distribution)
-    marginal_y: dict = get_marginal_distribution(inputs_y, joint_distribution)
+    marginal_xy: DiscreteDist = get_marginal_distribution(idxs_xy, joint_distribution)
+    marginal_y: DiscreteDist = get_marginal_distribution(idxs_y, joint_distribution)
 
     ptw: dict = {}
     avg: float = 0
+
     for state in marginal_xy.keys():
         if marginal_xy[state] > 0.0:
-
             p_y: float = marginal_y[state[Nx:]]
             h: float = -np.log2(marginal_xy[state] / p_y)
             ptw[((state[:Nx]), (state[Nx:]))] = h
@@ -89,9 +93,9 @@ def conditional_entropy(
 
 
 def mutual_information(
-    inputs_x: tuple,
-    inputs_y: tuple,
-    joint_distribution: dict,
+    idxs_x: tuple[int, ...],
+    idxs_y: tuple[int, ...],
+    joint_distribution: DiscreteDist,
 ) -> tuple[dict, float]:
     """
     Computes the mutual information between X and Y.
@@ -100,9 +104,9 @@ def mutual_information(
 
     Parameters
     ----------
-    inputs_x : tuple
+    idxs_x : tuple
         The indices of the X variable(s).
-    inputs_y : tuple
+    idxs_y : tuple
         The indices of the Y variable(s).
     joint_distribution : dict
         The joint probability distribution.
@@ -116,22 +120,22 @@ def mutual_information(
         The pointwise mutual information for each state in the joint distribution.
     avg : float
         The average mutual information
-    
+
     """
 
-    Nx: int = len(inputs_x)
+    Nx: int = len(idxs_x)
 
-    inputs_xy: tuple = inputs_x + inputs_y
[…]
