### 2025-11-03 4243b2e
Type-hinting neural estimators
