### 2025-11-03 9cad8d9
Type-hinting neural estimators
