+++ syntropy/neural/multivariate_mi.py
@@ -1,4 +1,3 @@
-import numpy as np
 import scipy.stats as stats
 import torch
 from shannon import differential_entropy
@@ -14,27 +13,30 @@ def total_correlation(
     verbose: bool = False,
 ) -> tuple[float, float]:
     """
+    Computes the total correlation of the data 
+
 
     Parameters
     ----------
-    idxs : tuple[int]
-
+    idxs : tuple[int, ...]
+        The tuple of indices the differential entropy is computed for.
     data : torch.Tensor
-
+        The training data, in samples x features format.
     context : None | tuple[int]
-
+        If not None, the indices of the conditioning variables.
     data_test : None | torch.Tensor
-
+        If not None, the testing data in samples x features format.
     flow_kwargs : dict
-
+        Arguments for the utils.initalize_flow function.
     train_kwargs : dict
-
+        Arguments for the utils.train_flow function.
     verbose : bool
+        Whether to print the training progress.
 
 
     Returns
     -------
-
+    float
 
 
     """
@@ -91,7 +93,37 @@ def higher_order_information(
     flow_kwargs: dict = None,
     train_kwargs: dict = None,
     verbose: bool = False,
-):
+) -> dict[str, float]:
+    """
+    Computes the O-information, S-information, total correlation, and dual total correlation for the data. 
+    Computing them all as a set is more efficient than computing each one independently.
+
+    Parameters
+    ----------
+    idxs : tuple[int, ...]
+        The tuple of indices the differential entropy is computed for.
+    data : torch.Tensor
+        The training data, in samples x features format.
+    context : None | tuple[int]
+        If not None, the indices of the conditioning variables.
+    data_test : None | torch.Tensor
+        If not None, the testing data in samples x features format.
+    flow_kwargs : dict
+        Arguments for the utils.initalize_flow function.
+    train_kwargs : dict
+        Arguments for the utils.train_flow function.
+    verbose : bool
+        Whether to print the training progress.
+        
+        
+
+    Returns
+    -------
+    dict[str, float]
+    
+        
+
+    """
     flow_kwargs = flow_kwargs or {}
     train_kwargs = train_kwargs or {}
 
@@ -100,8 +132,8 @@ def higher_order_information(
     else:
         context_arg = context
 
-    lookup_marginals: dict = {(i,): None for i in idxs}
-    lookup_residuals: dict = {tuple(j for j in idxs if j != i): None for i in idxs}
+    lookup_marginals: dict[tuple[int, ...], float] = {(i,): 0.0 for i in idxs}
+    lookup_residuals: dict[tuple[int, ...], float] = {tuple(j for j in idxs if j != i): 0.0 for i in idxs}
     
     N: int = len(idxs)
     
