+++ syntropy/__init__.py
@@ -1,3 +1,3 @@
 from . import discrete
 from . import gaussian
-from . import distribution
+from .discrete import distributions
