### 2025-11-04 fae5b81
KNN estimators (WIP)
