+++ syntropy/gaussian/multivariate_mi.py
@@ -4,7 +4,7 @@ from syntropy.gaussian.shannon import local_differential_entropy
 
 
 def local_total_correlation(
-    data: np.ndarray, cov: np.ndarray, inputs: tuple = (-1,)
+    data: np.ndarray, cov: np.ndarray, idxs: tuple = (-1,)
 ) -> np.ndarray:
     """
     The local total correlation.
@@ -33,22 +33,22 @@ def local_total_correlation(
         "The data and covariance matrix must have the same dimensionality"
     )
 
-    if inputs[0] == -1:
-        _inputs = tuple(i for i in range(data.shape[0]))
+    if idxs[0] == -1:
+        _idxs = tuple(i for i in range(data.shape[0]))
     else:
-        _inputs = inputs
-    N = len(_inputs)
+        _idxs = idxs
+    N = len(_idxs)
 
-    whole = local_differential_entropy(data[_inputs, :], cov[np.ix_(_inputs, _inputs)])
+    whole = local_differential_entropy(data[_idxs, :], cov[np.ix_(_idxs, _idxs)])
 
     sum_parts = np.zeros_like(whole)
     for i in range(N):
-        sum_parts += local_differential_entropy(data[_inputs[i], :])
+        sum_parts += local_differential_entropy(data[_idxs[i], :])
 
     return sum_parts - whole
 
 
-def total_correlation(cov: np.ndarray, inputs: tuple = (-1,)) -> float:
+def total_correlation(cov: np.ndarray, idxs: tuple = (-1,)) -> float:
     """
     The expected total correlation.
 
@@ -67,7 +67,7 @@ def total_correlation(cov: np.ndarray, inputs: tuple = (-1,)) -> float:
     ----------
     cov : np.ndarray
         The covariance matrix.
-    inputs : tuple, optional
+    idxs : tuple, optional
         The specific subset of variables to compute the total correlation of.
         Defaults to computing the TC of the entire covariance matrix.
 
@@ -76,12 +76,12 @@ def total_correlation(cov: np.ndarray, inputs: tuple = (-1,)) -> float:
     float
         The expected total correlation.
     """
-    if inputs[0] == -1:
-        _inputs = tuple(i for i in range(cov.shape[0]))
+    if idxs[0] == -1:
+        _idxs = tuple(i for i in range(cov.shape[0]))
     else:
-        _inputs = inputs
+        _idxs = idxs
 
-    _cov: np.ndarray = cov[np.ix_(_inputs, _inputs)]
+    _cov: np.ndarray = cov[np.ix_(_idxs, _idxs)]
 
     # Converting to a correlation/coherence matrix.
     diag: np.ndarray = np.sqrt(np.diag(_cov))
@@ -93,7 +93,7 @@ def total_correlation(cov: np.ndarray, inputs: tuple = (-1,)) -> float:
 
 
 def local_k_wms(
-    k: int, data: np.ndarray, cov: np.ndarray, inputs: tuple = (-1,)
+    k: int, data: np.ndarray, cov: np.ndarray, idxs: tuple = (-1,)
 ) -> np.ndarray:
     """
 
@@ -107,7 +107,7 @@ def local_k_wms(
     cov : np.ndarray, optional
         The covariance matrix that defines the distribution.
         If unspecified it is computed directly from the data.
-    inputs : tuple, optional
+    idxs : tuple, optional
         The specific subset of variables to compute the total correlation of.
         Defaults to computing the TC of the entire covariance matrix.
 
@@ -117,26 +117,26 @@ def local_k_wms(
         The series of local k_{wms}.
 
     """
-    if inputs[0] == -1:
-        _inputs: tuple = tuple(i for i in range(data.shape[0]))
+    if idxs[0] == -1:
[…]
