### 2025-11-04 442ba90
KNN estimators (WIP)
