+++ syntropy/gaussian/shannon.py
@@ -1,5 +1,6 @@
 import numpy as np
 import scipy.stats as stats
+from numpy.typing import NDArray
 
 H_SINGLE: float = np.log(np.sqrt(2.0 * np.pi * np.e))
 LN_TWO_PI_E: float = np.log(2.0 * np.pi * np.e)
@@ -9,7 +10,9 @@ SQRT_TWO_PI: float = np.sqrt(TWO_PI)
 COV_NULL = np.array([[-1]])
 
 
-def differential_entropy(cov: np.ndarray, inputs: tuple = (-1,)) -> float:
+def differential_entropy(
+    cov: NDArray[np.floating], idxs: tuple[int, ...] = (-1,)
+) -> float:
     """
     Computes the expected differential entropy of a multivariate
     distribution parameterized by a covariance matrix using a
@@ -17,9 +20,9 @@ def differential_entropy(cov: np.ndarray, inputs: tuple = (-1,)) -> float:
 
     Parameters
     ----------
-    cov : np.ndarray
+    cov : NDArray[np.floating]
         The covariance matrix that defines the distribution.
-    inputs : tuple, optional
+    idxs : tuple, optional
         The specific subset of variables to compute the total correlation of.
         Defaults to computing the TC of the entire covariance matrix.
 
@@ -29,34 +32,34 @@ def differential_entropy(cov: np.ndarray, inputs: tuple = (-1,)) -> float:
 
     """
 
-    if inputs[0] == -1:
+    if idxs[0] == -1:
         return stats.multivariate_normal(cov=cov, allow_singular=True).entropy()
     else:
-        if len(inputs) == 1:
+        if len(idxs) == 1:
             return H_SINGLE
         else:
             return stats.multivariate_normal(
-                cov=cov[inputs, :][:, inputs], allow_singular=True
+                cov=cov[idxs, :][:, idxs], allow_singular=True
             ).entropy()
 
 
 def local_differential_entropy(
-    data: np.ndarray, cov: np.ndarray = COV_NULL
-) -> np.ndarray:
+    data: NDArray[np.floating], cov: NDArray[np.floating] = COV_NULL
+) -> NDArray[np.floating]:
     """
     Computes the framewise differential entropy for a set of variables.
 
     Parameters
     ----------
-    data : np.ndarray
+    data : NDArray[np.floating]
         The data in channels x samples format.
-    cov : np.ndarray, optional
+    cov : NDArray[np.floating], optional
         The covariance matrix that defines the distribution.
         If none is provided, it is computed from the data object.
 
     Returns
     -------
-    np.ndarray
+    NDArray[np.floating]
         The series of pointwise entropies.
 
     """
@@ -80,7 +83,9 @@ def local_differential_entropy(
 
 
 def conditional_entropy(
-    idxs_x: tuple, idxs_y: tuple, cov: np.ndarray = COV_NULL
+    idxs_x: tuple[int, ...],
+    idxs_y: tuple[int, ...],
+    cov: NDArray[np.floating] = COV_NULL,
 ) -> float:
     """
     Computes the conditional entropy of X given Y.
@@ -91,7 +96,7 @@ def conditional_entropy(
         The indices of the variables to compute the conditional entropy on.
     idxs_y : tuple
         The indices of the conditioning set.
-    cov : np.ndarray, optional
+    cov : NDArray[np.floating], optional
         The covariance matrix that defines the distribution.
         If none is provided, it is computed from the data object.
 
@@ -110,8 +115,11 @@ def conditional_entropy(
 
 
 def local_conditional_entropy(
[…]
