### 2025-11-05 49f0646
KNN estimators (WIP)
