### 2025-11-05 d60d78b
KNN estimators (docs)
