+++ syntropy/knn/multivariate_mi.py
@@ -12,24 +12,39 @@ def total_correlation(
     algorithm: int = 1,
 ) -> tuple[NDArray[np.floating], float]:
     """
+    Computes the total correlation using either the first or second KSG estimator presented in 
 
+    Kraskov, A., Stögbauer, H., & Grassberger, P. (2004).
+    Estimating mutual information.
+    Physical Review E, 69(6), 066138.
+    https://doi.org/10.1103/PhysRevE.69.066138
+    
+    :math: `TC(X) = \\sum_{i=1}^{N}H(X_i) - H(X)`
+
+    See:
+        Watanabe, S. (1960). 
+        Information Theoretical Analysis of Multivariate Correlation.
+        IBM Journal of Research and Development, 4(1), Article 1.
+        https://doi.org/10.1147/rd.41.0066
+    
     Parameters
     ----------
-    algorithm :
-
-    idxs_x : tuple[int, ...]
-
-    idxs_y : tuple[int, ...]
-
-    k : int
-
     data : NDArray[np.floating]
-
+        Data array of shape (n_variables, n_samples)
+    k : int
+        Number of nearest neighbors
+    idxs : tuple[int, ...]
+        Indices of variables to use (-1 means all)
+    algorithm : int
+        Whether to use algorithm 1 or 2. 
+        Defaults to 1
 
     Returns
     -------
-    tuple[NDArray[np.floating], float]
-
+    NDArray[np.floating 
+        The local total correlation for each sample.
+    float
+        The expected total correlation over all samples
 
     """
     assert algorithm in {1, 2}, "Algorithm must be 1 or 2."
@@ -44,20 +59,30 @@ def total_correlation_1(
     data: NDArray[np.floating], k: int, idxs: tuple[int, ...] = (-1,)
 ) -> tuple[NDArray[np.floating], float]:
     """
+    Computes the Kraskov, Stogbauer, Grassberger estimate of the total correlation using the first algorithm presented in:
+    
+    Kraskov, A., Stögbauer, H., & Grassberger, P. (2004).
+    Estimating mutual information.
+    Physical Review E, 69(6), 066138.
+    https://doi.org/10.1103/PhysRevE.69.066138
+    
+    :math: `TC(X) = \psi(k) - (m-1)\psi(N) - \rangle \psi(n_{x_{1}}+1) + \ldots + \psi(n_{x_{N}}+1)\langle`
 
     Parameters
     ----------
     data : NDArray[np.floating]
-
+        Data array of shape (n_variables, n_samples)
     k : int
-
+        Number of nearest neighbors
     idxs : tuple[int, ...]
-
+        Indices of variables to use (-1 means all)
 
     Returns
     -------
-    tuple[NDArray[np.floating], float]
-
+    NDArray[np.floating 
+        The local total correlation for each sample.
+    float
+        The expected total correlation over all samples
 
     """
     idxs_: tuple[int,...] = check_idxs(idxs, data.shape[0])
@@ -90,26 +115,31 @@ def total_correlation_2(
 ) -> tuple[NDArray[np.floating], float]:
     """
     Computes the Kraskov, Stogbauer, Grassberger estimate of the total correlation using the second algorithm presented in:
-
+    
     Kraskov, A., Stögbauer, H., & Grassberger, P. (2004).
     Estimating mutual information.
     Physical Review E, 69(6), 066138.
     https://doi.org/10.1103/PhysRevE.69.066138
+    
+    :math: `TC(X) = \psi(k) - ((m-1)/k) - (m-1)\psi(N) - \langle \psi(n_{x_{1}}) + \ldots + \psi(n_{x_{N}}) \rangle`
 
     Parameters
     ----------
     data : NDArray[np.floating]
[…]
