### 2025-11-05 37126af
KNN estimators (docs)
