### 2025-11-05 88ee2af
Type-hinting Gaussians
