### 2025-11-05 b6ac633
Type-hinting Gaussians
