+++ syntropy/gaussian/multivariate_mi.py
@@ -1,11 +1,12 @@
 import numpy as np
 from syntropy.gaussian.utils import COV_NULL
 from syntropy.gaussian.shannon import local_differential_entropy
+from numpy.typing import NDArray
 
 
 def local_total_correlation(
-    data: np.ndarray, cov: np.ndarray, idxs: tuple = (-1,)
-) -> np.ndarray:
+    data: NDArray[np.floating], cov: NDArray[np.floating], idxs: tuple[int, ...] = (-1,)
+) -> NDArray[np.floating]:
     """
     The local total correlation.
 
@@ -19,13 +20,13 @@ def local_total_correlation(
 
     Parameters
     ----------
-    data : np.ndarray
+    data : NDArray[np.floating]
         The data in channels x samples format.
     inptuts: tuple
         The indices of the channels to include.
     Returns
     -------
-    np.ndarray
+    NDArray[np.floating]
         The local total correaltion for each frame.
 
     """
@@ -48,9 +49,13 @@ def local_total_correlation(
     return sum_parts - whole
 
 
-def total_correlation(cov: np.ndarray, idxs: tuple = (-1,)) -> float:
+def total_correlation(
+    cov: NDArray[np.floating], idxs: tuple[int, ...] = (-1,)
+) -> float:
     """
     The expected total correlation.
+
+    :math: TC(X) = \\sum_{i=1}^{N}H(X_i) - H(X)
 
     See:
         Watanabe, S. (1960). Information Theoretical Analysis of Multivariate Correlation.
@@ -65,7 +70,7 @@ def total_correlation(cov: np.ndarray, idxs: tuple = (-1,)) -> float:
 
     Parameters
     ----------
-    cov : np.ndarray
+    cov : NDArray[np.floating]
         The covariance matrix.
     idxs : tuple, optional
         The specific subset of variables to compute the total correlation of.
@@ -81,30 +86,38 @@ def total_correlation(cov: np.ndarray, idxs: tuple = (-1,)) -> float:
     else:
         _idxs = idxs
 
-    _cov: np.ndarray = cov[np.ix_(_idxs, _idxs)]
+    _cov: NDArray[np.floating] = cov[np.ix_(_idxs, _idxs)]
 
     # Converting to a correlation/coherence matrix.
-    diag: np.ndarray = np.sqrt(np.diag(_cov))
-    d_inv: np.ndarray = np.diag(1.0 / diag)
+    diag: NDArray[np.floating] = np.sqrt(np.diag(_cov))
+    d_inv: NDArray[np.floating] = np.diag(1.0 / diag)
 
-    corr: np.ndarray = d_inv @ _cov @ d_inv
+    corr: NDArray[np.floating] = d_inv @ _cov @ d_inv
 
     return -np.linalg.slogdet(corr)[1] / 2
 
 
 def local_k_wms(
-    k: int, data: np.ndarray, cov: np.ndarray, idxs: tuple = (-1,)
-) -> np.ndarray:
+    k: int,
+    data: NDArray[np.floating],
+    cov: NDArray[np.floating],
+    idxs: tuple[int, ...] = (-1,),
+) -> NDArray[np.floating]:
     """
+    A utility function that computes the local generalized form
+    of the O-information, S-information, and DTC.
+
+    :math:`k_{WMS}(x) = (N-k)tc(x) - \\sum_{i=1}^{N} tc(x^{-i})`
+
 
     Parameters
     ----------
     k : int
         The integer value that defines whether one is computing
         S-info, DTC, or negative O-information.
-    data : np.ndarray
+    data : NDArray[np.floating]
         The data in channels x samples format.
-    cov : np.ndarray, optional
+    cov : NDArray[np.floating], optional
[…]
