### 2025-11-05 0998971
Type-hinting Gaussians Fixing bug in Gaussian k_wms rate.
