+++ syntropy/gaussian/temporal.py
@@ -1,18 +1,19 @@
 import numpy as np
 import scipy.signal as signal
 import scipy.integrate as integrate
+from numpy.typing import NDArray
 
 
 def construct_csd_tensor(
-    idxs: tuple, data: np.ndarray, fs: int = 1, nperseg: int = 1024
-) -> tuple[np.ndarray, np.ndarray]:
+    idxs: tuple[int, ...], data: NDArray[np.floating], fs: int = 1, nperseg: int = 1024
+) -> tuple[NDArray[np.floating], NDArray[np.floating]]:
     """
 
     Parameters
     ----------
-    idxs : tuple
+    idxs : tuple[int, ...]
 
-    data : np.ndarray
+    data : NDArray[np.floating]
 
     fs : int
 
@@ -21,12 +22,12 @@ def construct_csd_tensor(
 
     Returns
     -------
-    np.ndarray
+    NDArray[np.floating]
 
 
     """
     N: int = len(idxs)
-    S: np.ndarray = np.zeros((nperseg, N, N))
+    S: NDArray[np.floating] = np.zeros((nperseg, N, N))
 
     for i in range(N):
         for j in range(i + 1):
@@ -48,11 +49,11 @@ def construct_csd_tensor(
 
 
 def differential_entropy_rate(
-    idxs: tuple,
-    data: np.ndarray,
+    idxs: tuple[int, ...],
+    data: NDArray[np.floating],
     fs: int = 1,
     nperseg: int = 1024,
-) -> tuple[np.ndarray, float]:
+) -> tuple[NDArray[np.floating], float]:
     """
     Computes the differential entropy rate of a potentially multivariate stochastic process.
 
@@ -60,9 +61,9 @@ def differential_entropy_rate(
 
     Parameters
     ----------
-    idxs : tuple
+    idxs : tuple[int, ...]
         The indices of the channels to include in the analysis.
-    data : np.ndarray
+    data : NDArray[np.floating]
         The data in channels x samples format.
     fs: int
         The sampling rate of the time series data in Hz.
@@ -74,7 +75,7 @@ def differential_entropy_rate(
 
     Returns
     -------
-    np.ndarray
+    NDArray[np.floating]
         The local entropies for each frequency band.
     float
         The average entropy across the whole spectrum.
@@ -93,12 +94,12 @@ def differential_entropy_rate(
 
 
 def mutual_information_rate(
-    idxs_x: tuple,
-    idxs_y: tuple,
-    data: np.ndarray,
+    idxs_x: tuple[int, ...],
+    idxs_y: tuple[int, ...],
+    data: NDArray[np.floating],
     fs: int = 1,
     nperseg: int = 1024,
-) -> tuple[np.ndarray, float]:
+) -> tuple[NDArray[np.floating], float]:
     """
     Computes the mutual information rate between two (potentially multivariate) Gaussian processes.
 
@@ -117,11 +118,11 @@ def mutual_information_rate(
 
     Parameters
     ----------
-    idxs_x : tuple
+    idxs_x : tuple[int, ...]
         The indices of the channels to include in the inputs.
-    idxs_y : tuple
+    idxs_y : tuple[int, ...]
         The indices of the channels to include in the target.
-    data : np.ndarray
+    data : NDArray[np.floating]
         The data in channels x samples format.
     fs: int
         The sampling rate of the time series data in Hz.
@@ -133,7 +134,7 @@ def mutual_information_rate(
 
     Returns
     -------
-    np.ndarray
+    NDArray[np.floating]
         The local mutual informations for each frequency band.
     float
         The average mutual information across the whole spectrum.
[…]
