+++ syntropy/nonparametric/shannon.py
@@ -1,11 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Fri Mar  7 18:11:42 2025
-
-@author: thosvarley
-"""
-
-import numpy as np 
-from scipy.special import digamma
-from scipy.spatial import cKDTree
