+++ syntropy/knn/shannon.py
@@ -1,16 +1,8 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Fri Mar  7 18:11:42 2025
-
-@author: thosvarley
-"""
-
 import numpy as np
 from numpy.typing import NDArray
 from scipy.special import digamma
 from scipy.spatial import cKDTree
-from utils import check_idxs, build_tree_and_get_distances, get_counts_from_tree
+from .utils import check_idxs, build_tree_and_get_distances, get_counts_from_tree
 
 
 def differential_entropy(
@@ -19,26 +11,38 @@ def differential_entropy(
     """
     Computes the differential entropy using the Kozachenko-Leoneko estimator.
 
-    :math: `H(X) = -\psi(k)+\psi(N) + (1/N)\\sum_{i=1}^{N}\log d_i`
+    .. math::
+        \hat{H}(X) = -\psi(k)+\psi(N) + (1/N)\\sum_{i=1}^{N}\log d_i
 
     Parameters
     ----------
     data : NDArray[np.floating]
-        Data array of shape (n_variables, n_samples)
+        Numpy array of shape (n_variables, n_samples)
     k : int
         Number of nearest neighbors
     idxs : tuple[int, ...]
         Indices of variables to use (-1 means all)
-:
+
     Returns
     -------
-    NDArray[np.floating
+    NDArray[np.floating]
         The local differential entropy for each sample.
     float
         The expected differential entropy over all samples
 
+    References
+    ----------
+    Delattre, S., & Fournier, N. (2017).
+    On the Kozachenko–Leonenko entropy estimator.
+    Journal of Statistical Planning and Inference, 185, 69–93.
+    https://doi.org/10.1016/j.jspi.2017.01.004
+
+    Kozachenko, L. F., & Leonenko, N. N. (1987).
+    Sample Estimate of the Entropy of a~Random Vector.
+    Problems of Information Transmission, 23(2), 9.
 
     """
+
     idxs_: tuple[int, ...] = check_idxs(idxs, data.shape[0])
 
     d: int = len(idxs_)
@@ -62,6 +66,12 @@ def mutual_information(
     algorithm: int = 1,
 ) -> tuple[NDArray[np.floating], float]:
     """
+    A wrapper function for the two KSG mutual information functions.
+
+    See Also
+    --------
+    mutual_information_1 : Using the KSG-1 algorithm.
+    mutual_information_2 : Using the KSG-2 algorithm.
 
     Parameters
     ----------
@@ -72,7 +82,7 @@ def mutual_information(
     k : int
         Number of nearest neighbors
     data : NDArray[np.floating]
-        Data array of shape (n_variables, n_samples)
+        Numpy array of shape (n_variables, n_samples)
     algorithm : int
         Whether to use algorithm 1 or 2.
         Defaults to 1
@@ -85,6 +95,7 @@ def mutual_information(
         The expected mutual information over all samples
 
     """
+
     assert algorithm in {1, 2}, "Algorithm must be 1 or 2."
 
     if algorithm == 1:
@@ -98,12 +109,10 @@ def mutual_information_1(
 ) -> tuple[NDArray[np.floating], float]:
     """
     Computes the Kraskov, Stogbauer, Grassberger estimate of the bivariate mutual information
-    using the first algorithm presented in:
+    using the first algorithm presented in Kraskov et. al., (2004)
 
-    Kraskov, A., Stögbauer, H., & Grassberger, P. (2004).
-    Estimating mutual information.
-    Physical Review E, 69(6), 066138.
-    https://doi.org/10.1103/PhysRevE.69.066138
+    .. math::
[…]
