+++ syntropy/neural/__init__.py
@@ -1,0 +1,16 @@
+from .shannon import (
+    differential_entropy,
+    mutual_information,
+)
+
+from .multivariate_mi import (
+    total_correlation,
+    higher_order_information
+)
+
+__all__ = [
+    "differential_entropy",
+    "mutual_information",
+    "total_correlation",
+    "higher_order_information"
+]
