+++ docs/source/index.rst
@@ -1,0 +1,17 @@
+.. Syntropy documentation master file, created by
+   sphinx-quickstart on Wed Nov  5 16:27:52 2025.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Syntropy documentation
+======================
+
+Add your content using ``reStructuredText`` syntax. See the
+`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
+documentation for details.
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+

[…]
