+++ syntropy/__init__.py
@@ -1,3 +1,0 @@
-from . import discrete
-from . import gaussian
-from .discrete import distributions
