### 2025-11-06 13f0047
readthedocs configuration
