+++ docs/source/conf.py
@@ -50,3 +50,6 @@ napoleon_use_rtype = True
 napoleon_preprocess_types = False
 napoleon_type_aliases = None
 napoleon_attr_annotations = True
+
+# Typehints settings
+autodoc_typehints = "description"
