### 2025-11-06 98d5673
Readthedocs config
