### 2025-11-06 d910323
Readthedocs config
