### 2025-11-06 63b244a
Fix bug in local_differential_entropy
