### 2025-11-06 ad2a327
Fix bug in local_differential_entropy
