### 2025-11-06 bd3e187
Type hinting Gaussian estimators
