+++ syntropy/gaussian/utils.py
@@ -34,7 +34,7 @@ LAYERS_4: dict = {
     for val in set(PATHS_4.values())
 }
 
-COV_NULL = np.array([[-1]])
+COV_NULL: NDArray[np.floating] = np.array([[-1.0]])
 
 # %% LIBRARY
 
@@ -54,7 +54,8 @@ def make_powerset(iterable):
     """
     Computes the powerset of a collection of elements.
 
-    :math:`\\mathcal{P}(\\{X_1,X_2,X_3\\}) \\to (\\{\\}, \\{X_1\\}, \\{X_2\\}, \\{X_3\\}, \\{X_1,X_2\\}, \\{X_1,X_3\\}, \\{X_1,X_2,X_3\\} )`
+    .. math:: 
+        \\mathcal{P}(\\{X_1,X_2,X_3\\}) \\to (\\{\\}, \\{X_1\\}, \\{X_2\\}, \\{X_3\\}, \\{X_1,X_2\\}, \\{X_1,X_3\\}, \\{X_1,X_2,X_3\\} )
 
     """
     xs: list = list(iterable)
@@ -81,9 +82,9 @@ def local_precompute_sources(data: NDArray[np.floating], cov: NDArray[np.floatin
 
     Parameters
     ----------
-    data : np.ndarray
+    data : NDArray[np.floating]
         The data in channels x samples format.
-    cov : np.ndarray, optional
+    cov : NDArray[np.floating], optional
         The covariance matrix of the data.
         The default is COV_NULL.
 
@@ -108,8 +109,8 @@ def local_precompute_sources(data: NDArray[np.floating], cov: NDArray[np.floatin
 
 
 def hmin_differential_redundancy(
-    atom: tuple, sources: dict, data: np.ndarray, cov: np.ndarray = COV_NULL
-) -> np.ndarray:
+    atom: tuple, sources: dict, data: NDArray[np.floating], cov: NDArray[np.floating] = COV_NULL
+) -> NDArray[np.floating]:
     """
     For a collection of sources :math:`\\alpha=\\{a_1, a_2, \\ldots, a_k\\}`, computes
     the redundnat entropy shared by all sources as:
@@ -128,17 +129,17 @@ def hmin_differential_redundancy(
         The partial information atom. In the form :math:`((a_1,),(a_2,)\\ldots)`.
     sources : dict
         The pre-computed collection of sources.
-    data : np.ndarray
+    data : NDArray[np.floating]
         The data, assumed to be in sources x samples format.
 
     Returns
     -------
-    np.ndarray
+    NDArray[np.floating]
         The local redundancies for each sample.
 
 
     """
-    h_min: np.ndarray = np.repeat(np.inf, data.shape[1])
+    h_min: NDArray[np.floating] = np.repeat(np.inf, data.shape[1])
     source: tuple = tuple()
 
     for source in atom:
@@ -153,9 +154,9 @@ def imin_differential_redundancy(
     sources: dict,
     inputs: tuple,
     target: tuple,
-    data: np.ndarray,
-    cov: np.ndarray = COV_NULL,
-) -> np.ndarray:
+    data: NDArray[np.floating],
+    cov: NDArray[np.floating] = COV_NULL,
+) -> NDArray[np.floating]:
     """
     Computes the differential redundnacy between a partial information atom :math:`\\alpha=\\{a_i,\\ldots,a_k\\}` and a (potentially multivariate) target. Uses a Gaussian estimator for the local entropies.
 
@@ -180,12 +181,12 @@ def imin_differential_redundancy(
         The indicies of the inputs - one index per element of the tuple.
     target : tuple
         The (potentially multivariate) indices of the collective target.
-    data : np.ndarray
+    data : NDArray[np.floating]
[…]
