+++ syntropy/gaussian/temporal.py
@@ -57,7 +57,8 @@ def differential_entropy_rate(
     """
     Computes the differential entropy rate of a potentially multivariate stochastic process.
 
-    :math:`H(X)=\\frac{1}{4\\pi} \int_{-\\pi}^{\\pi} \\log \\left( (2\\pi e)^N |S_X(\\omega)| \\right) \\, d\\omega`
+    .. math::
+        H(X)=\\frac{1}{4\\pi} \int_{-\\pi}^{\\pi} \\log \\left( (2\\pi e)^N |S_X(\\omega)| \\right) \\, d\\omega
 
     Parameters
     ----------
@@ -103,18 +104,8 @@ def mutual_information_rate(
     """
     Computes the mutual information rate between two (potentially multivariate) Gaussian processes.
 
-    :math:`I(X; Y) = \\frac{1}{4\\pi} \\int_{-\\pi}^{\\pi} \\log \\left( \\frac{  |S_X(\\omega)||S_Y(\\omega)| }{ |S_{XY}(\\omega)| } \\right) d\\omega`
-
-    See:
-
-    Faes, L., Sparacino, L., Mijatovic, G., Antonacci, Y., Ricci, L., Marinazzo, D., & Stramaglia, S. (2025).
-    Partial Information Rate Decomposition (No. arXiv:2502.04550). arXiv.
-    https://doi.org/10.48550/arXiv.2502.04550
-
-    Faes, L., Pernice, R., Mijatovic, G., Antonacci, Y., Krohova, J. C., Javorka, M., & Porta, A. (2021).
-    Information decomposition in the frequency domain: A new framework to study cardiovascular and cardiorespiratory oscillations.
-    Philosophical Transactions of the Royal Society A: Mathematical, Physical and Engineering Sciences, 379(2212), 20200250.
-    https://doi.org/10.1098/rsta.2020.0250
+    .. math:: 
+        I(X; Y) = \\frac{1}{4\\pi} \\int_{-\\pi}^{\\pi} \\log \\left( \\frac{  |S_X(\\omega)||S_Y(\\omega)| }{ |S_{XY}(\\omega)| } \\right) d\\omega
 
     Parameters
     ----------
@@ -138,6 +129,18 @@ def mutual_information_rate(
         The local mutual informations for each frequency band.
     float
         The average mutual information across the whole spectrum.
+    
+    References
+    ----------
+    Faes, L., Sparacino, L., Mijatovic, G., Antonacci, Y., Ricci, L., Marinazzo, D., & Stramaglia, S. (2025).
+    Partial Information Rate Decomposition (No. arXiv:2502.04550). arXiv.
+    https://doi.org/10.48550/arXiv.2502.04550
+
+    Faes, L., Pernice, R., Mijatovic, G., Antonacci, Y., Krohova, J. C., Javorka, M., & Porta, A. (2021).
+    Information decomposition in the frequency domain: A new framework to study cardiovascular and cardiorespiratory oscillations.
+    Philosophical Transactions of the Royal Society A: Mathematical, Physical and Engineering Sciences, 379(2212), 20200250.
+    https://doi.org/10.1098/rsta.2020.0250
+
     """
     Nx: int = len(idxs_x)
     Ny: int = len(idxs_y)
@@ -172,7 +175,8 @@ def total_correlation_rate(
     """
     A straightforward extension of the mutual information rate to the total correlation.
 
-    :math:`TC(X,Y,\\ldots,Z) = \\frac{1}{4\\pi} \\int_{-\\pi}^{\\pi} \\log \\left( \\frac{  |S_X(\\omega)||S_Y(\\omega)|\\ldots|S_Z(\\omega)| }{ |S_{XY\\ldots Z}(\\omega)| } \\right) d\\omega`
+    .. math:: 
[…]
