+++ syntropy/discrete/decompositions.py
@@ -1,33 +1,25 @@
 import numpy as np
 from .utils import mobius_inversion
+from typing import Callable
 
 
 def partial_information_decomposition(
     redundancy: str,
     inputs: tuple,
     target: tuple,
-    joint_distribution: dict[tuple, float],
+    joint_distribution: dict[tuple[int, ...], float],
 ) -> (dict, dict):
     """
     Computes the partial information decomposition for up to four input variables
     onto one (potentially joint) target variable.
 
-    The available redundancy functions are i_min from Finn and Lizier
-    and i_sx from Makkeh et al.. See:
+    The available redundancy functions are :math:`i_{\\min}` from Finn and Lizier and :math:`i_{sx}` from Makkeh et al..
 
-        Finn, C., & Lizier, J. T. (2018).
-        Pointwise Partial Information Decomposition Using
-        the Specificity and Ambiguity Lattices.
-        Entropy, 20(4), Article 4.
-        https://doi.org/10.3390/e20040297
+    .. math:: 
 
-        Makkeh, A., Gutknecht, A. J., & Wibral, M. (2021).
-        Introducing a differentiable measure of pointwise
-        shared information.
-        Physical Review E, 103(3), 032149.
-        https://doi.org/10.1103/PhysRevE.103.032149
-
-
+        i_{\\min}(\\alpha;t) &= \\min h(\\alpha_i) - \min h(\\alpha_i|t) \\\\
+           i_{sx}(\\alpha;t) &= \\log\\frac{P(t)-P(t\\cap(\\alpha_1\\cup ...\\cup\\alpha_k))}{1-P(\\bar\\alpha_1\\cap ...\\cap\\alpha_N)}
+    
     Parameters
     ----------
     redundancy : str
@@ -52,6 +44,25 @@ def partial_information_decomposition(
     avg : dict
         The expected value for each partial information atom.
 
+    References
+    ----------
+    Williams, P. L., & Beer, R. D. (2010). 
+    Nonnegative Decomposition of Multivariate Information. 
+    arXiv:1004.2515 [Math-Ph, Physics:Physics, q-Bio]. 
+    http://arxiv.org/abs/1004.2515
+
+    Finn, C., & Lizier, J. T. (2018).
+    Pointwise Partial Information Decomposition Using
+    the Specificity and Ambiguity Lattices.
+    Entropy, 20(4), Article 4.
+    https://doi.org/10.3390/e20040297
+
+    Makkeh, A., Gutknecht, A. J., & Wibral, M. (2021).
+    Introducing a differentiable measure of pointwise
+    shared information.
+    Physical Review E, 103(3), 032149.
+    https://doi.org/10.1103/PhysRevE.103.032149
+    
     """
 
     ptw, avg = mobius_inversion(
@@ -66,27 +77,18 @@ def partial_information_decomposition(
 
 
 def partial_entropy_decomposition(
-    redundancy: str, joint_distribution: dict[tuple, float]
+    redundancy: str, joint_distribution: dict[tuple[int, ...], float]
 ) -> (dict, dict):
     """
     Computes the partial entropy decomposition of a joint distribution
     with up to four elements.
 
     The available redundancy functions are h_min from Finn and Lizier
-    and h_sx from Varley et al., See:
-
-        Finn, C., & Lizier, J. T. (2020).
-        Generalised Measures of Multivariate Information Content.
-        Entropy, 22(2), Article 2.
-        https://doi.org/10.3390/e22020216
-
-        Varley, T. F., Pope, M., Maria Grazia, P., Joshua, F., & Sporns, O. (2023).
-        Partial entropy decomposition reveals higher-order
-        information structures in human brain activity.
[…]
