+++ syntropy/discrete/shannon.py
@@ -10,7 +10,7 @@ def shannon_entropy(joint_distribution: DiscreteDist) -> tuple[dict, float]:
     Computes the Shannon entropy of the distribution :math:`P(x)`.
 
     .. math::
-        H(X) = -\\sum_{x} P(x) \\log P(x)
+        H(X) = -\\sum_{x\\in\\mathcal{X}} P(x) \\log P(x)
 
     To compute the entropy of a subset of the variables in the joint
     distribution, use the :func:`syntropy.discrete.utils.get_marginals` function from the utils
@@ -48,7 +48,8 @@ def conditional_entropy(
     """
     Computes the conditional entropy of X given Y.
 
-    :math:`H(X|Y) = H(X,Y) - H(Y)`
+    .. math::
+        H(X|Y) = H(X,Y) - H(Y)
 
     Parameters
     ----------
@@ -100,7 +101,11 @@ def mutual_information(
     """
     Computes the mutual information between X and Y.
 
-    :math:`I(X;Y) = H(X) + H(Y) - H(X,Y)`
+    .. math:: 
+        I(X;Y) &= H(X) + H(Y) - H(X,Y) \\\\
+               &= H(X) - H(X|Y) \\\\
+               &= H(Y) - H(Y|X) \\\\
+               &= H(X,Y) - H(X|Y) - H(Y|X)
 
     Parameters
     ----------
@@ -154,7 +159,9 @@ def conditional_mutual_information(
     """
     Computes the mutual information between X and Y condioned on Z.
 
-    :math:`I(X,Y|Z) = H(X|Z) + H(Y|Z) - H(X,Y|Z)`
+    .. math:: 
+        I(X,Y|Z) &= H(X|Z) + H(Y|Z) - H(X,Y|Z) \\\\
+                 &= I(X;Y,Z) - I(X;Z)
 
     Parameters
     ----------
@@ -208,7 +215,8 @@ def kullback_leibler_divergence(
     Computes the Kullback-Leibler divergence from a prior distribution P(X) and
     and posterior distribution Q(X).
 
-    :math:`D_{KL}(P||Q) = \\sum_{x} P(x) \\log \\frac{P(x)}{Q(x)}`
+    .. math:: 
+        D_{KL}(P||Q) = \\sum_{x} P(x) \\log \\frac{P(x)}{Q(x)}
 
     Parameters
     ----------

[…]
