+++ syntropy/discrete/multivariate_mi.py
@@ -13,20 +13,18 @@ from .optimization import constrained_maximum_entropy_distributions
 binom_lookup = {N: {k: comb(N, k, exact=True) for k in range(N)} for N in range(16)}
 
 
-def total_correlation(joint_distribution) -> (tuple[dict, float], float):
+def total_correlation(joint_distribution: dict[tuple[int, ...], float]) -> (tuple[dict, float], float):
     """
     Computes the average and pointwise total correlations:
 
-    :math:`TC(X) = D_{KL}(P(X) || \\prod_{i=1}^{N}P(X_i)`
+    .. math::
 
-    See:
-        Watanabe, S. (1960). Information Theoretical Analysis of Multivariate Correlation.
-        IBM Journal of Research and Development, 4(1), Article 1.
-        https://doi.org/10.1147/rd.41.0066
+        TC(X) &= D_{KL}(P(X) || \\prod_{i=1}^{N}P(X_i) \\\\
+              &= \\sum_{i=1}^{N}H(X_i) - H(X)
 
     Parameters
     ----------
-    joint_distribution : dict[tuple, float]
+    joint_distribution : dict[tuple[int, ...]], float]
         The joint probability distribution.
         Keys are tuples corresponding to the state of each element.
         The valules are the probabilities.
@@ -37,6 +35,18 @@ def total_correlation(joint_distribution) -> (tuple[dict, float], float):
         The pointwise TC .
     avg : float
         The average TC.
+    
+    References
+    ----------
+    Watanabe, S. (1960). Information Theoretical Analysis of Multivariate Correlation.
+    IBM Journal of Research and Development, 4(1), Article 1.
+    https://doi.org/10.1147/rd.41.0066
+
+    Tononi, G., Sporns, O., & Edelman, G. M. (1994). 
+    A measure for brain complexity: Relating functional segregation and integration in the nervous system. 
+    Proceedings of the National Academy of Sciences, 91(11), Article 11. 
+    https://doi.org/10.1073/pnas.91.11.5033
+    
     """
 
     maxent_distribution = constrained_maximum_entropy_distributions(
@@ -48,37 +58,40 @@ def total_correlation(joint_distribution) -> (tuple[dict, float], float):
     return ptw, avg
 
 
-def k_wms(k: int, joint_distribution: dict[tuple, float]) -> (tuple[dict, float], float):
+def k_wms(k: int, joint_distribution: dict[tuple[int, ...], float]) -> (tuple[dict, float], float):
     """
     S-information, DTC, and negative O-information can all be written in a general form:
 
-    :math:`WMS^{k}(X) = (N-k)TC(X) - \\sum_{i=1}^{N}TC(X^{-i})`
+    .. math::
 
-    See:
-        Varley, T. F., Pope, M., Faskowitz, J., & Sporns, O. (2023).
-        Multivariate information theory uncovers synergistic subsystems of the human cerebral cortex.
-        Communications Biology, 6(1), Article 1.
-        https://doi.org/10.1038/s42003-023-04843-w
+        WMS^{k}(X) = (N-k)TC(X) - \\sum_{i=1}^{N}TC(X^{-i})
 
     Parameters
     ----------
     k : int
-        DESCRIPTION.
-    joint_distribution : dict[tuple, float]
+        The scale parameter
+    joint_distribution : dict[tuple[int, ...], float]
         The joint probability distribution.
         Keys are tuples corresponding to the state of each element.
         The valules are the probabilities.
 
     Returns
     -------
-    ptw : dict[tuple, float]
+    ptw : dict[tuple[int, ...], float]
         The pointwise w-WMS .
     avg : float
         The average k-WMS.
 
+    References
+    ----------
[…]
