+++ syntropy/gaussian/multivariate_mi.py
@@ -7,7 +7,10 @@ def local_total_correlation(
     data: NDArray[np.floating], cov: NDArray[np.floating], idxs: tuple[int, ...] = (-1,)
 ) -> NDArray[np.floating]:
     """
-    The local total correlation.
+    The local total correlation. Note that this measure can be negative.
+
+    .. math:: 
+        tc(x) = \\sum_{i=1}^{N}h(x_i) - h(x)
 
     Parameters
     ----------
@@ -60,8 +63,17 @@ def total_correlation(
     """
     The expected total correlation.
 
-    ..math:: 
-        `TC(X) = \\sum_{i=1}^{N}H(X_i) - H(X)`
+    .. math::
+
+        TC(X) &= D_{KL}(P(X) || \\prod_{i=1}^{N}P(X_i) \\\\
+              &= \\sum_{i=1}^{N}H(X_i) - H(X)
+
+    For Gaussian random variables, the estimator is:
+
+        .. math::
+            \\hat{TC}(X) = \\frac{-\log R}{2}
+
+    Where :math:`R` is the Pearson correlation matrix.
 
     Parameters
     ----------
@@ -161,11 +173,11 @@ def local_k_wms(
 
 def k_wms(k: int, cov: NDArray[np.floating], idxs: tuple[int, ...] = (-1,)) -> float:
     """
-    A utility function that computes the generalized form
-    of the O-information, S-information, and DTC.
+    S-information, DTC, and negative O-information can all be written in a general form:
 
-    .. math:: 
-        K_{WMS}(X) = (N-k)TC(X) - \\sum_{i=1}^{N} TC(X^{-i})
+    .. math::
+
+        WMS^{k}(X) = (N-k)TC(X) - \\sum_{i=1}^{N}TC(X^{-i})
 
     Parameters
     ----------
@@ -208,6 +220,11 @@ def local_s_information(
 ) -> NDArray[np.floating]:
     """
     Compute local S-information using Gaussian estimation.
+    
+    .. math::
+        \\sigma(X) &= \\sum_{i=1}^{N}i(x_i;x^{-i}) \\\\
+                   &= N\\times tc(x) - \\sum_{i=1}^{N}tc(x^{-i}) \\\\
+                   &= tc(x) + dtc(x)
 
     Parameters
     ----------
@@ -237,6 +254,12 @@ def s_information(cov: NDArray[np.floating], idxs: tuple[int, ...] = (-1,)) -> f
     """
     Compute S-information using Gaussian estimation.
 
+    .. math:: 
+
+        \\Sigma(X) &= \\sum_{i=1}^{N}I(X_i;X^{-i}) \\\\
+                   &= N\\times TC(X) - \\sum_{i=1}^{N}TC(X^{-i}) \\\\
+                   &= TC(X) + DTC(X)
+    
     Parameters
     ----------
     data : NDArray[np.floating]
@@ -276,7 +299,12 @@ def local_dual_total_correlation(
     idxs: tuple[int, ...] = (-1,),
 ) -> NDArray[np.floating]:
     """
-    Computes the dual total correlation using Gaussian estimation. 
+    Computes the dual total correlation using Gaussian estimation. Note that this measure can be negative.
+    
+    .. math:: 
+
+        dtc(x) &= h(x) - \\sum_{i=1}^{N}h(x_i|x^{-i}) \\\\
+               &= (N-1)\\times tc(x) - \\sum_{i=1}^{N}tc(x^{-i})
 
     Parameters
     ----------
@@ -306,7 +334,12 @@ def dual_total_correlation(
     cov: NDArray[np.floating], idxs: tuple[int, ...] = (-1,)
 ) -> float:
     """
-    Computes the dual total correlation. 
+    Computes the dual total correlation using Gaussian estimation.
+
[…]
