+++ docs/source/index.rst
@@ -78,7 +78,7 @@ Example
    examples
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 3
    :caption: API Reference:
 
    api/syntropy.discrete
