+++ .readthedocs.yaml
@@ -19,5 +19,7 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
   install:
+    - method: pip
+      path: .
   - requirements: docs/requirements.txt
 
