+++ .readthedocs.yaml
@@ -21,5 +21,5 @@ python:
   install:
     - method: pip
       path: .
-  - requirements: docs/requirements.txt
+    - requirements: docs/requirements.txt
 
