### 2025-11-07 7e7c4c7
KSG conditional MI
