+++ docs/build/html/_sources/api/syntropy.gaussian.rst.txt
@@ -9,42 +9,10 @@ syntropy.gaussian package
 Submodules
 ----------
 
-syntropy.gaussian.decompositions module
----------------------------------------
-
-.. automodule:: syntropy.gaussian.decompositions
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-syntropy.gaussian.multivariate\_mi module
------------------------------------------
-
-.. automodule:: syntropy.gaussian.multivariate_mi
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
 syntropy.gaussian.optimization module
 -------------------------------------
 
 .. automodule:: syntropy.gaussian.optimization
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-syntropy.gaussian.shannon module
---------------------------------
-
-.. automodule:: syntropy.gaussian.shannon
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-syntropy.gaussian.temporal module
----------------------------------
-
-.. automodule:: syntropy.gaussian.temporal
    :members:
    :show-inheritance:
    :undoc-members:
