+++ docs/build/html/_sources/index.rst.txt
@@ -78,7 +78,7 @@ Example
    examples
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 1
    :caption: API Reference:
 
    api/syntropy.discrete
