### 2025-11-07 05ee5e3
KSG conditional MI
