+++ syntropy/neural/shannon.py
@@ -1,20 +1,19 @@
 import torch
 import nflows
 from .utils import initialize_flow, train_flow, evaluate_flow
+from typing import Any
 
 
 def differential_entropy(
     idxs: tuple[int, ...],
     data: torch.Tensor,
-    context: None | tuple[int, ...] = None,
     data_test: None | torch.Tensor = None,
-    flow_kwargs: dict = None,
-    train_kwargs: dict = None,
+    flow_kwargs: None | dict[str, Any] = None,
+    train_kwargs: None | dict[str, Any] = None,
     verbose: bool = False,
-) -> tuple[float, nflows.flows.base.Flow]:
-
+) -> tuple[torch.Tensor, float, nflows.flows.base.Flow]:
     """
-    Computes the differential entropy of the data. 
+    Computes the differential entropy of the data.
 
     Parameters
     ----------
@@ -22,8 +21,6 @@ def differential_entropy(
         The tuple of indices the differential entropy is computed for.
     data : torch.Tensor
         The training data, in samples x features format.
-    context : None | tuple[int]
-        If not None, the indices of the conditioning variables.
     data_test : None | torch.Tensor
         If not None, the testing data in samples x features format.
     flow_kwargs : dict
@@ -32,24 +29,18 @@ def differential_entropy(
         Arguments for the utils.train_flow function.
     verbose : bool
         Whether to print the training progress.
-        
+
 
     Returns
     -------
     float
 
     nflows.flows.base.Flow
-        
+
 
     """
     flow_kwargs = flow_kwargs or {}
     train_kwargs = train_kwargs or {}
-
-    # select context tensor if provided
-    context_arg: None | torch.Tensor = None
-    if context is not None:
-        context_arg = data[:, context]
-        flow_kwargs["dim_context"] = len(context)
 
     # initialize the flow
     dim: int = len(idxs)
@@ -59,37 +50,96 @@ def differential_entropy(
     flow = train_flow(
         flow=flow,
         data=data[:, idxs],
-        context=context_arg,
         verbose=verbose,
         **train_kwargs,
     )
     # evaluation
     eval_data = data if data_test is None else data_test
-    eval_context = (
-        None
-        if context is None
-        else (data_test[:, context] if data_test is not None else context_arg)
+    ptw, avg = evaluate_flow(flow, eval_data[:, idxs])
+
+    return ptw, avg, flow
+
+
+def conditional_entropy(
+    idxs_x: tuple[int, ...],
+    idxs_y: tuple[int, ...],
+    data: torch.Tensor,
+    data_test: None | torch.Tensor = None,
+    flow_kwargs: None | dict[str, Any] = None,
+    train_kwargs: None | dict[str, Any] = None,
+    verbose: bool = False,
+) -> tuple[torch.Tensor, float, nflows.flows.base.Flow]:
+    """
+    Computes the differential entropy of the data.
+
+    Parameters
+    ----------
+    idxs_x : tuple[int, ...]
+        The tuple of indices the differential entropy is computed for.
+    idxs_y : tuple[int, ...]
+        The tuple of indices to condition on.
+    data : torch.Tensor
+        The training data, in samples x features format.
+    data_test : None | torch.Tensor
+        If not None, the testing data in samples x features format.
+    flow_kwargs : dict
+        Arguments for the utils.initalize_flow function.
+    train_kwargs : dict
+        Arguments for the utils.train_flow function.
+    verbose : bool
[…]
