+++ syntropy/knn/shannon.py
@@ -43,15 +43,15 @@ def differential_entropy(
 
     """
 
-    idxs_: tuple[int, ...] = check_idxs(idxs, data.shape[0])
+    idxs: tuple[int, ...] = check_idxs(idxs, data.shape[0])
 
-    d: int = len(idxs_)
+    d: int = len(idxs)
     N: int = data.shape[1]
 
     psi_k: float = digamma(k)
     psi_N: float = digamma(N)
 
-    _, distances, _ = build_tree_and_get_distances(data[idxs_], k=k)
+    _, distances, _ = build_tree_and_get_distances(data[idxs], k=k)
 
     ptw: NDArray[np.floating] = -psi_k + psi_N + d * np.log(2 * distances[:, -1])
 
@@ -223,5 +223,228 @@ def mutual_information_2(
         tree, distances, _ = build_tree_and_get_distances(data_idx.T, k=k)
         counts = get_counts_from_tree(tree, data_idx.T, eps)
         ptw -= digamma(counts)
+
+    return ptw, ptw.mean()
+
+
+def conditional_mutual_information_1(
+    idxs_x: tuple[int, ...],
+    idxs_y: tuple[int, ...],
+    idxs_z: tuple[int, ...],
+    k: int,
+    data: NDArray[np.floating],
+) -> tuple[NDArray[np.floating], float]:
+    """
+    Computes the conditional mutual information I(X;Y|Z) using the KSG algorithm 1
+    as described in Frenzel & Pompe (2007).
+
+    The conditional mutual information is estimated using:
+
+    .. math::
+        \\hat{I}(X;Y|Z) = \\psi(k) - \\langle \\psi(n_{xz}+1) + \\psi(n_{yz}+1) - \\psi(n_z+1) \\rangle
+
+    where n_{xz}, n_{yz}, and n_z are the counts of neighbors within the epsilon-ball
+    in the respective subspaces.
+
+    Parameters
+    ----------
+    idxs_x : tuple[int, ...]
+        Indices of the x-variable.
+    idxs_y : tuple[int, ...]
+        Indices of the y-variable.
+    idxs_z : tuple[int, ...]
+        Indices of the conditioning variable z.
+    k : int
+        Number of nearest neighbors
+    data : NDArray[np.floating]
+        Numpy array of shape (n_variables, n_samples)
+
+    Returns
+    -------
+    NDArray[np.floating]
+        The local conditional mutual information for each sample.
+    float
+        The expected conditional mutual information over all samples
+
+    References
+    ----------
+    Frenzel, S., & Pompe, B. (2007).
+    Partial Mutual Information for Coupling Analysis of Multivariate Time Series.
+    Physical Review Letters, 99(20), 204101.
+    https://doi.org/10.1103/PhysRevLett.99.204101
+
+    Kraskov, A., Stögbauer, H., & Grassberger, P. (2004).
+    Estimating mutual information.
+    Physical Review E, 69(6), 066138.
+    https://doi.org/10.1103/PhysRevE.69.066138
+
+    """
+
+    idxs_x = check_idxs(idxs_x, data.shape[0])
+    idxs_y = check_idxs(idxs_y, data.shape[0])
+    idxs_z = check_idxs(idxs_z, data.shape[0])
+
+    # Combine indices for joint space
+    idxs_xyz: tuple[int, ...] = idxs_x + idxs_y + idxs_z
+    idxs_xz: tuple[int, ...] = idxs_x + idxs_z
+    idxs_yz: tuple[int, ...] = idxs_y + idxs_z
+
+    N: int = data.shape[1]
+    psi_k: float = digamma(k)
+
[…]
