+++ syntropy/gaussian/utils.py
@@ -1,38 +1,7 @@
 import numpy as np
-import networkx as nx
 import itertools as it
 from numpy.typing import NDArray
 
-from ..lattices import LATTICE_2, LATTICE_3, LATTICE_4
-
-Atom = tuple[tuple[int, ...], ...]
-BOTTOM_2: Atom = ((0,), (1,))
-BOTTOM_3: Atom = ((0,), (1,), (2,))
-BOTTOM_4: Atom = ((0,), (1,), (2,), (3,))
-
-PATHS_2: dict = nx.shortest_paths.shortest_path_length(
-    LATTICE_2, source=None, target=BOTTOM_2
-)
-LAYERS_2: dict[int, int] = {
-    val: {key for key in PATHS_2.keys() if PATHS_2[key] == val}
-    for val in set(PATHS_2.values())
-}
-
-PATHS_3: dict[int, int] = nx.shortest_paths.shortest_path_length(
-    LATTICE_3, source=None, target=BOTTOM_3
-)
-LAYERS_3: dict[int, int] = {
-    val: {key for key in PATHS_3.keys() if PATHS_3[key] == val}
-    for val in set(PATHS_3.values())
-}
-
-PATHS_4: dict = nx.shortest_paths.shortest_path_length(
-    LATTICE_4, source=None, target=BOTTOM_4
-)
-LAYERS_4: dict = {
-    val: {key for key in PATHS_4.keys() if PATHS_4[key] == val}
-    for val in set(PATHS_4.values())
-}
 
 COV_NULL: NDArray[np.floating] = np.array([[-1.0]])
 
@@ -63,271 +32,6 @@ def make_powerset(iterable):
     return it.chain.from_iterable(it.combinations(xs, n) for n in range(len(xs) + 1))
 
 
-def unpack_atom(atom: tuple) -> set:
-    """
-    A utitlity function to unpack tuples.
-    """
-    varset: set = set()
-    for s in {*atom}:
-        varset.update(set(s))
-
-    return varset
-
-
-def local_precompute_sources(data: NDArray[np.floating], cov: NDArray[np.floating] = COV_NULL):
-    """
-    A utility function that pre-computes the local entropies
-    of every subset of the data. This speeds up the PID/GID/PED
-    by orders of magnitude.
-
-    Parameters
-    ----------
-    data : NDArray[np.floating]
-        The data in channels x samples format.
-    cov : NDArray[np.floating], optional
-        The covariance matrix of the data.
-        The default is COV_NULL.
-
-    Returns
-    -------
-    dict
-        A dictionary of the local entropies for every source.
-
-    """
-    if cov[0][0] == -1:
-        cov = np.cov(data, ddof=0.0)
-
-    N: int = data.shape[0]
-    joint: tuple = tuple(i for i in range(N))
-    sources: list = list(make_powerset(joint))
-    sources.remove(())
-
-    return {
-        source: local_differential_entropy(data[source, :], cov[source, :][:, source])
-        for source in sources
-    }
-
-
-def hmin_differential_redundancy(
-    atom: tuple, sources: dict, data: NDArray[np.floating], cov: NDArray[np.floating] = COV_NULL
-) -> NDArray[np.floating]:
-    """
-    For a collection of sources :math:`\\alpha=\\{a_1, a_2, \\ldots, a_k\\}`, computes
-    the redundnat entropy shared by all sources as:
-
-    :math:`h_{\\cap}^{min}(\\alpha) = \\min_{i}(a_{i})`
-
-    See:
-        Finn, C., & Lizier, J. T. (2020).
-        Generalised Measures of Multivariate Information Content.
-        Entropy, 22(2), Article 2.
-        https://doi.org/10.3390/e22020216
-
-    Parameters
-    ----------
[…]
