+++ syntropy/discrete/decompositions.py
@@ -1,8 +1,502 @@
 import numpy as np
-from .utils import mobius_inversion
+import networkx as nx
 from typing import Callable
+from .utils import make_powerset, reduce_state
+
+from ..lattices import LATTICE_2, LATTICE_3, LATTICE_4
+
+BOTTOM_2: tuple = ((0,), (1,))
+BOTTOM_3: tuple = ((0,), (1,), (2,))
+BOTTOM_4: tuple = ((0,), (1,), (2,), (3,))
+
+PATHS_2: dict = nx.shortest_paths.shortest_path_length(
+    LATTICE_2, source=None, target=BOTTOM_2
+)
+LAYERS_2: dict = {
+    val: {key for key in PATHS_2.keys() if PATHS_2[key] == val}
+    for val in set(PATHS_2.values())
+}
+
+PATHS_3: dict = nx.shortest_paths.shortest_path_length(
+    LATTICE_3, source=None, target=BOTTOM_3
+)
+LAYERS_3: dict = {
+    val: {key for key in PATHS_3.keys() if PATHS_3[key] == val}
+    for val in set(PATHS_3.values())
+}
+
+PATHS_4: dict = nx.shortest_paths.shortest_path_length(
+    LATTICE_4, source=None, target=BOTTOM_4
+)
+LAYERS_4: dict = {
+    val: {key for key in PATHS_4.keys() if PATHS_4[key] == val}
+    for val in set(PATHS_4.values())
+}
+
+def local_precompute_sources(joint_distribution: dict[tuple, float]) -> dict:
+    """
+    A utility function that computes the local entropy of each subset of
+    elements. This speeds up the computation using the hmin function
+    considerably,
+
+    Parameters
+    ----------
+    joint_distribution: dict[tuple, float]
+        The joint probability distribution.
+        Keys are tuples corresponding to the state of each element.
+        The valules are the probabilities.
+
+    Returns
+    -------
+    dict
+        For each state, the joint entropy of every subset.
+
+    """
+
+    N: int = len(list(joint_distribution.keys())[0])
+
+    sources: list = list(make_powerset(range(N)))
+    sources.remove(())
+
+    local_entropies: dict = {
+        state: {source: {} for source in sources} for state in joint_distribution.keys()
+    }
+
+    state: tuple
+    for state in local_entropies.keys():
+
+        source: tuple
+        for source in local_entropies[state]:
+
+            # For a given state x, and a given source s, computes the total probability mass of the joint
+            # distribution consistent with the state of the source.
+            # E.g. if state = (0,0,0) and source = (0,1), probability_mass = P(X0=0 AND X1=0)
+            probability_mass = sum(
+                [
+                    joint_distribution[key]
+                    for key in joint_distribution.keys()
+                    if reduce_state(key, source) == reduce_state(state, source)
+                ]
+            )
+
+            if probability_mass > 0:
+                local_entropies[state][source] = -np.log2(probability_mass)
+            else:
+                local_entropies[state][
+                    source
+                ] = 0  # Set log2(0) to 0 since impossible events contain no information.
+
+    return local_entropies
 
 
+def hmin_discrete_redundancy(
+    atom: tuple, state: tuple, sources: dict, joint_distribution: dict[tuple, float]
+) -> float:
+    """
+    For a collection of sources :math:`\\alpha=\\{a_1, a_2, \\ldots, a_k\\}`, computes
+    the redundnat entropy shared by all sources as:
+
+
+    :math:`h_{\\cap}^{min}(\\alpha) = \\min_{i}h(a_i)`
+
+    See:
+        Finn, C., & Lizier, J. T. (2020).
+        Generalised Measures of Multivariate Information Content.
+        Entropy, 22(2), Article 2.
[…]
