### 2025-11-12 #8934 @thosvarley
@cbueth Here's the readthedocs (I'll add the link to the README asap)
https://syntropy.readthedocs.io/en/latest/

I was hoping to put it on PyPI once I'd made it through peer review.

All the best
~Thomas
