### 2025-11-12 242dd36
Refactoring connected information
