+++ syntropy/discrete/multivariate_mi.py
@@ -13,7 +13,9 @@ from .optimization import constrained_maximum_entropy_distributions
 binom_lookup = {N: {k: comb(N, k, exact=True) for k in range(N)} for N in range(16)}
 
 
-def total_correlation(joint_distribution: dict[tuple[int, ...], float]) -> (tuple[dict, float], float):
+def total_correlation(
+    joint_distribution: dict[tuple[int, ...], float],
+) -> (tuple[dict, float], float):
     """
     Computes the average and pointwise total correlations:
 
@@ -58,7 +60,9 @@ def total_correlation(joint_distribution: dict[tuple[int, ...], float]) -> (tupl
     return ptw, avg
 
 
-def k_wms(k: int, joint_distribution: dict[tuple[int, ...], float]) -> (tuple[dict, float], float):
+def k_wms(
+    k: int, joint_distribution: dict[tuple[int, ...], float]
+) -> (tuple[dict, float], float):
     """
     S-information, DTC, and negative O-information can all be written in a general form:
 
@@ -91,7 +95,7 @@ def k_wms(k: int, joint_distribution: dict[tuple[int, ...], float]) -> (tuple[di
 
     """
 
-    states: list[tuple[int,...]] = list(joint_distribution.keys())
+    states: list[tuple[int, ...]] = list(joint_distribution.keys())
     N: int = len(states[0])
 
     ptw_tc: dict
@@ -106,7 +110,6 @@ def k_wms(k: int, joint_distribution: dict[tuple[int, ...], float]) -> (tuple[di
     ptw_sum_parts = {state: 0 for state in states}
 
     for i in range(N):
-
         residuals: tuple = tuple(j for j in range(N) if j != i)
 
         reduced_distribution: dict = marginalize_out((i,), joint_distribution)
@@ -129,7 +132,9 @@ def k_wms(k: int, joint_distribution: dict[tuple[int, ...], float]) -> (tuple[di
     return ptw, avg
 
 
-def s_information(joint_distribution: dict[tuple[int, ...], float]) -> (tuple[dict, float], float):
+def s_information(
+    joint_distribution: dict[tuple[int, ...], float],
+) -> (tuple[dict, float], float):
     """
     Computes the local and expected S-information for the joint distribution.
 
@@ -174,7 +179,9 @@ def s_information(joint_distribution: dict[tuple[int, ...], float]) -> (tuple[di
     return ptw, avg
 
 
-def dual_total_correlation(joint_distribution: dict[tuple[int, ...], float]) -> (tuple[dict, float], float):
+def dual_total_correlation(
+    joint_distribution: dict[tuple[int, ...], float],
+) -> (tuple[dict, float], float):
     """
     Computes the local and expected dual total correlations for the joint distribution.
 
@@ -217,7 +224,9 @@ def dual_total_correlation(joint_distribution: dict[tuple[int, ...], float]) -> 
     return ptw, avg
 
 
-def o_information(joint_distribution: dict[tuple[int, ...], float]) -> (tuple[dict, float], float):
+def o_information(
+    joint_distribution: dict[tuple[int, ...], float],
+) -> (tuple[dict, float], float):
     """
     Computes the local and expected O-informations for the joint distribution.
     O-information quantifies the balance between redundancy (positive values) and synergy (negative values) in multivariate information.
@@ -262,11 +271,13 @@ def o_information(joint_distribution: dict[tuple[int, ...], float]) -> (tuple[di
     return {state: -ptw[state] for state in ptw.keys()}, -avg
 
 
-def co_information(joint_distribution: dict[tuple[int, ...], float]) -> (tuple[dict, float], float):
+def co_information(
+    joint_distribution: dict[tuple[int, ...], float],
+) -> (tuple[dict, float], float):
     """
[…]
