### 2025-11-12 #9407 @editorialbot
[REVIEW]: "Syntropy": a Python package for discrete and continuous information theory.
**Submitting author:** @thosvarley (<a href="http://orcid.org/0000-0002-3317-9882">Thomas F. Varley</a>)
**Repository:** https://github.com/thosvarley/syntropy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
