### 2025-11-12 thosvarley/syntropy#3 @cbueth
JOSS review
Hello @thosvarley, in this issue I will write my questions and points of improvements for the review https://github.com/openjournals/joss-reviews/issues/9407#issuecomment-3523215280. Before going through the paper and testing the code, some code chores:

- You might want to populate the `.gitignore` with relevant points:
  - `__pycache__/` is not essential to the package (also at `syntropy/__pycache__`)
  - Is there a reason you want to keep `docs/build`? In usual package structures they are not needed and can be built from sources, including readthedocs.

  See an example of often ignored files and folders with python at https://github.com/github/gitignore/blob/main/Python.gitignore. You might want something like:

```.gitignore
# Byte-compiled / optimized / DLL files
__pycache__/
*.py[cod]
*$py.class

# Development
*.egg-info
docs/build/
dist/

# Jupyter Notebook
.ipynb_checkpoints
```

- I understand you only want to offer APIs at subpackage levels? It is recommended that `syntropy/__init__.py` has at least some information about the package e.g. `__version__` and `__author__`. Often this is where a user-friendly API is centralised.
- Just out of curiosity. Do you use any formatting following any style guides?
- You might want to extend the walkthrough for the Neural estimators https://syntropy.readthedocs.io/en/latest/quickstart.html#neural-estimators.
- Do you plan to add more tests? It seems there are only tests for the discrete and gaussian case. It would be better to have at least all core features testes, better to have all functionality tested. What is often done:
  - Using the same file structure in `tests/` as in the package's folder and having a test file for each source file.
  - Measuring the covered lines with https://coverage.readthedocs.io/en/latest/. A coverage of 50%+ is desirable and usually easily attainable.
- Nice you are using a `pyproject.toml`. I saw it still misses the project URLs. These will appear on the PyPI page, see https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#urls.
- For the `dependencies`, are there any specific version limitations to the dependent packages? I see there are none set; please assure there are no mayor limitations for older versions, or decide for lower limits. You can consult the package histories or changelogs.
- It is great you have a https://syntropy.readthedocs.io/en/latest/contributing.html#contributing page, I do think it still needs to address some questions. It should answer how to contribute, where and anything to remember, e.g. testing. What is always useful to reference there is a code of conduct ([CODE_OF_CONDUCT.md](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file)). It is also very useful to include it at `CONTRIBUTING.md`, then github will automatically display it properly and note it on the repository page https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors.
- Maybe you want to reconsider the package name, as it already has been taken on PyPI https://pypi.org/project/syntropy/. Alternatively, it is possible to have a different pypi/pip name, to the package name itself, but it can be confusing to new users.

These are all points you can consider at your own disgression. Do not feel pressured to implement all of them, but I would be happy to get your input to each of them and hope the package profits off of them. If anything is unclear, please write a short comment and I'm happy to help 👍
