### 2025-11-12 #9407 @cbueth
Before starting my checklist, I did a shallow code quality review, find my comment in https://github.com/thosvarley/syntropy/issues/3#issue-3617788022. To my impression the package has a mature core. The packaging still has a few rough edges, and testing needs to be augmented.
