### 2025-11-12 thosvarley/syntropy#3 @thosvarley
Thanks for all of the suggestions - I should note that I'm a neuroscientist by training, not a software engineer, so this is my first foray into the open source space. I can go through these point by point, but in general the answer to a question like "why didn't you do [standard thing]" is probably that I have no idea that [standard thing] was standard. Or even a thing.
