### 2025-11-13 thosvarley/syntropy#3 @cbueth
Okay, thanks for the heads-up, that sounds exciting! I hope my comments were formulated not too assuming. They should all be suggestions for directions to look into, and you decide what to do with it. I will not suggest anything too unnecessary. The list might seem much, but I think you will go through most of them rather quick. So if anything is unclear, or you want further input, just ask. As Chris said, the JOSS review approach is a supportive one. As far as to what I have seen, the software quality makes a good impression to me.
