### 2025-11-18 c729680
Mixed datatype estimators
