+++ syntropy/mixed/__init__.py
@@ -1,0 +1,11 @@
+from .shannon import (
+        shannon_entropy, 
+        conditional_entropy, 
+        mutual_information
+        )
+
+__all__ = [
+        "shannon_entropy",
+        "conditional_entropy",
+        "mutual_information"
+        ]
