### 2025-11-18 61f0501
Mixed datatype estimators
