### 2025-11-18 e8a882e
Type hinting
