+++ syntropy/mixed/shannon.py
@@ -5,41 +5,49 @@ from .gaussian import differential_entropy, local_differential_entropy
 
 
 def shannon_entropy(
-    discrete_vars: NDArray[np.integer], continuous_vars: NDArray[np.floating]
+    discrete_vars: NDArray[np.integer],
+    continuous_vars: NDArray[np.floating],
 ) -> tuple[NDArray[np.floating], float]:
     """
+    For discrete :math:`X` and continuous :math:`Y`, leverages the identity .. math::
+        H(X,Y) = H(X) + H(Y | X)
 
     Parameters
     ----------
     discrete_vars : NDArray[np.integer]
-
+        Numpy array of the discrete variables, of shape (n_variables, n_samples)
     continuous_vars : NDArray[np.floating]
-
+        Numpy array of the continuous variables, of shape (n_variables, n_samples)
 
     Returns
     -------
-    tuple[NDArray[np.floating], float]
-
+    NDArray[np.floating]
+        The local entropy for each sample.
+    floating
+        The expected entropy over all samples.
 
     """
     assert discrete_vars.shape[1] == continuous_vars.shape[1], (
         "The discrete and continuous variables must have the same number of samples."
     )
 
+    unq: NDArray[np.integer]
+    counts: NDArray[np.integer]
     unq, counts = np.unique(discrete_vars, axis=1, return_counts=True)
-    probs = counts / counts.sum()
 
-    avg = -np.sum(probs * np.log(probs))
-    ptw = np.zeros(continuous_vars.shape[1])
+    probs: NDArray[np.floating] = counts / counts.sum()
 
-    conditional_entropy = 0.0
+    avg: float = -np.sum(probs * np.log(probs))
+    ptw: NDArray[np.floating] = np.zeros(continuous_vars.shape[1])
+
+    conditional_entropy: float = 0.0
     for i in range(unq.shape[1]):
-        state = unq[:, [i]]
-        mask = np.all(discrete_vars == state, axis=0)
+        state: NDArray[np.integer] = unq[:, [i]]
+        mask: NDArray[np.bool_] = np.all(discrete_vars == state, axis=0)
 
         ptw[mask] -= np.log(probs[i])
 
-        cov_conditional = np.cov(continuous_vars[:, mask], ddof=0)
+        cov_conditional: NDArray[np.floating] = np.cov(continuous_vars[:, mask], ddof=0)
 
         ptw[mask] += local_differential_entropy(
             continuous_vars[:, mask], cov_conditional
@@ -57,26 +65,27 @@ def conditional_entropy(
     conditional: str = "discrete",
 ) -> tuple[NDArray[np.floating], float]:
     """
-    For a discrete, potetially multivariate, random variable Y and a continuous, potentially multivariate random variable X, this estimator exploits the identity:
+    For discrete :math:`X` and continuous :math:`Y`, leverages the identity .. math::
+        H(X,Y) = H(X) + H(Y | X)
 
-    .. math::
-        H(X,Y) = H(Y|X) + H(Y)
-    
-    Where :math:'H(Y)' can be computed with the usual :math:'-\\sum P(x)\\log P(x)'
+    Either the discrete variables or the continuous variables can be conditioning variable (in which case the other is conditioned).
 
     Parameters
     ----------
     discrete_vars : NDArray[np.integer]
-
+        Numpy array of the discrete variables, of shape (n_variables, n_samples)
     continuous_vars : NDArray[np.floating]
-
+        Numpy array of the continuous variables, of shape (n_variables, n_samples)
     conditional : str
+        Wheter to condition the discrete variables on the continuous, or vice versa.
 
 
     Returns
     -------
-    tuple[NDArray[np.floating], float]
-
[…]
