### 2025-11-18 ecab9e3
Mixed datatype estimators
