### 2025-11-18 de4a345
Mixed datatype estimators
