+++ syntropy/mixed/shannon.py
@@ -1,12 +1,13 @@
 import numpy as np
 from numpy.typing import NDArray
-
-from .gaussian import differential_entropy, local_differential_entropy
-
+from .. import gaussian
+from .. import knn
 
 def shannon_entropy(
     discrete_vars: NDArray[np.integer],
     continuous_vars: NDArray[np.floating],
+    continuous_estimator: str = "gaussian",
+    k: int = 5,
 ) -> tuple[NDArray[np.floating], float]:
     """
     For discrete :math:`X` and continuous :math:`Y`, leverages the identity .. math::
@@ -18,6 +19,10 @@ def shannon_entropy(
         Numpy array of the discrete variables, of shape (n_variables, n_samples)
     continuous_vars : NDArray[np.floating]
         Numpy array of the continuous variables, of shape (n_variables, n_samples)
+    continuous_estimator : str 
+        Whether to use a Gaussian or KNN-based estimator of the continuous entropy. Options are "gaussian" or "knn".
+    k : int (optional)
+        If a KNN-based estimator is selected, sets the k-value. Defaults to 5.    
 
     Returns
     -------
@@ -38,21 +43,21 @@ def shannon_entropy(
     probs: NDArray[np.floating] = counts / counts.sum()
 
     avg: float = -np.sum(probs * np.log(probs))
-    ptw: NDArray[np.floating] = np.zeros(continuous_vars.shape[1])
+    ptw: NDArray[np.floating] = np.zeros((1,continuous_vars.shape[1]))
 
     conditional_entropy: float = 0.0
     for i in range(unq.shape[1]):
         state: NDArray[np.integer] = unq[:, [i]]
         mask: NDArray[np.bool_] = np.all(discrete_vars == state, axis=0)
 
-        ptw[mask] -= np.log(probs[i])
+        ptw[:, mask] -= np.log(probs[i])
 
         cov_conditional: NDArray[np.floating] = np.cov(continuous_vars[:, mask], ddof=0)
 
-        ptw[mask] += local_differential_entropy(
+        ptw[:,mask] += gaussian.local_differential_entropy(
             continuous_vars[:, mask], cov_conditional
         )
-        conditional_entropy += probs[i] * differential_entropy(cov_conditional)
+        conditional_entropy += probs[i] * gaussian.differential_entropy(cov_conditional)
 
     avg += conditional_entropy
 
@@ -105,23 +110,24 @@ def conditional_entropy(
 
     if conditional == "continuous":
         cov: NDArray[np.floating] = np.cov(continuous_vars, ddof=0)
-        ptw_marginal: NDArray[np.floating] = local_differential_entropy(
+        ptw_marginal: NDArray[np.floating] = gaussian.local_differential_entropy(
             continuous_vars, cov=cov
         )
-        avg_marginal: float = differential_entropy(cov)
+        avg_marginal: float = gaussian.differential_entropy(cov)
     elif conditional == "discrete":
         unq: NDArray[np.integer]
         counts: NDArray[np.integer]
         unq, counts = np.unique(discrete_vars, axis=1, return_counts=True)
         probs: NDArray[np.floating] = counts / counts.sum()
-        ptw_marginal: NDArray[np.floating] = np.zeros(discrete_vars.shape[1])
+        
+        ptw_marginal: NDArray[np.floating] = np.zeros((1,discrete_vars.shape[1]))
         avg_marginal: float = -np.sum(probs * np.log(probs))
 
         for i in range(unq.shape[1]):
             state: NDArray[np.integer] = unq[:, [i]]
             mask: NDArray[np.bool_] = np.all(discrete_vars == state, axis=0)
 
[…]
