### 2025-11-19 92fcf7e
KNN estimator dims
