+++ syntropy/knn/multivariate_mi.py
@@ -2,7 +2,7 @@ import numpy as np
 from numpy.typing import NDArray
 from scipy.special import digamma
 from scipy.spatial import cKDTree
-from .utils import check_idxs, build_tree_and_get_distances, get_counts_from_tree 
+from .utils import check_idxs, build_tree_and_get_distances, get_counts_from_tree
 
 
 def total_correlation(
@@ -12,7 +12,7 @@ def total_correlation(
     algorithm: int = 1,
 ) -> tuple[NDArray[np.floating], float]:
     """
-    A wrapper function for the two TC functions. 
+    A wrapper function for the two TC functions.
 
     Parameters
     ----------
@@ -23,7 +23,7 @@ def total_correlation(
     idxs : tuple[int, ...]
         Indices of variables to use (-1 means all)
     algorithm : int
-        Whether to use algorithm 1 or 2. 
+        Whether to use algorithm 1 or 2.
         Defaults to 1
 
     See Also
@@ -33,13 +33,13 @@ def total_correlation(
 
     Returns
     -------
-    NDArray[np.floating 
+    NDArray[np.floating
         The local total correlation for each sample.
     float
         The expected total correlation over all samples
 
     """
-    
+
     assert algorithm in {1, 2}, "Algorithm must be 1 or 2."
 
     if algorithm == 1:
@@ -52,9 +52,9 @@ def total_correlation_1(
     data: NDArray[np.floating], k: int, idxs: tuple[int, ...] = (-1,)
 ) -> tuple[NDArray[np.floating], float]:
     """
-    Computes the Kraskov, Stogbauer, Grassberger estimate of the total correlation using the first algorithm presented Kraskov et. al (2004) 
-    
-    .. math:: 
+    Computes the Kraskov, Stogbauer, Grassberger estimate of the total correlation using the first algorithm presented Kraskov et. al (2004)
+
+    .. math::
         \hat{TC}(X) = \psi(k) - (m-1)\psi(N) -\\langle \psi(n_{x_{1}}+1) + \ldots + \psi(n_{x_{N}}+1)\\rangle
 
     Parameters
@@ -68,7 +68,7 @@ def total_correlation_1(
 
     Returns
     -------
-    NDArray[np.floating 
+    NDArray[np.floating
         The local total correlation for each sample.
     float
         The expected total correlation over all samples
@@ -79,18 +79,18 @@ def total_correlation_1(
     Estimating mutual information.
     Physical Review E, 69(6), 066138.
     https://doi.org/10.1103/PhysRevE.69.066138
-    
-    Watanabe, S. (1960). 
+
+    Watanabe, S. (1960).
     Information Theoretical Analysis of Multivariate Correlation.
     IBM Journal of Research and Development, 4(1), Article 1.
     https://doi.org/10.1147/rd.41.0066
-    
+
     """
-    idxs_: tuple[int,...] = check_idxs(idxs, data.shape[0])
+    idxs_: tuple[int, ...] = check_idxs(idxs, data.shape[0])
 
     N: int = data.shape[1]
     m: int = len(idxs_)
-    
+
     psi_k, psi_N = digamma([k, N])
 
     tree: cKDTree = cKDTree(data[idxs_, :].T)
@@ -98,17 +98,17 @@ def total_correlation_1(
     distances, _ = tree.query(data[idxs_, :].T, k=k + 1, p=np.inf)
     eps: NDArray[np.floating] = distances[:, -1]
 
-    ptw: NDArray[np.floating] = np.zeros(N)
+    ptw: NDArray[np.floating] = np.zeros((1, N))
 
     for i in idxs_:
[…]
