### 2025-11-19 abdc6a9
Fixing broken KNN CMI estimator
