### 2025-11-19 f0f8df6
Fixing broken KNN CMI estimator
