### 2025-11-19 9f6ae6a
Debugging KNN estimators
