+++ syntropy/knn/__init__.py
@@ -1,22 +1,22 @@
 from .shannon import (
     differential_entropy,
     mutual_information,
-    conditional_mutual_information
+    #conditional_mutual_information
 )
 
-from .multivariate_mi import (
-    total_correlation,
-    dual_total_correlation,
-    s_information,
-    o_information
-)
-
+# from .multivariate_mi import (
+#     total_correlation,
+#     dual_total_correlation,
+#     s_information,
+#     o_information
+# )
+#
 __all__ = [
     "differential_entropy",
     "mutual_information",
-    "conditional_mutual_information",
-    "total_correlation",
-    "dual_total_correlation",
-    "s_information",
-    "o_information",
+    #"conditional_mutual_information",
+    # "total_correlation",
+    # "dual_total_correlation",
+    # "s_information",
+    # "o_information",
 ]
