### 2025-11-19 740d373
Tests for KNN estimators
