+++ syntropy/knn/__init__.py
@@ -4,18 +4,18 @@ from .shannon import (
     #conditional_mutual_information
 )
 
-# from .multivariate_mi import (
-#     total_correlation,
+from.multivariate_mi import (
+     total_correlation,
 #     dual_total_correlation,
 #     s_information,
 #     o_information
-# )
-#
+)
+
 __all__ = [
     "differential_entropy",
     "mutual_information",
     #"conditional_mutual_information",
-    # "total_correlation",
+     "total_correlation",
     # "dual_total_correlation",
     # "s_information",
     # "o_information",
