### 2025-11-19 46da4d6
KNN DTC
