+++ tests/test_knn.py
@@ -3,7 +3,7 @@ import pathlib
 import numpy as np
 import pandas as pd
 
-from syntropy.knn import mutual_information, total_correlation
+from syntropy.knn import mutual_information, total_correlation, dual_total_correlation
 
 data_path = pathlib.Path(__file__).parent
 data = pd.read_csv(data_path / "../examples/bold.csv", header=None).values
@@ -18,3 +18,9 @@ def test_total_correlation():
     mi2 = mutual_information(idxs_x=(0,), idxs_y=(1,), data=data, k=5, algorithm=2)
 
     assert tc1[1] == pytest.approx(mi1[1], abs=pytest_abs)
+
+def test_dual_total_correlation():
+    dtc1 = dual_total_correlation(data=data, k=5, idxs=(0, 1))
+    mi1 = mutual_information(idxs_x=(0,), idxs_y=(1,), data=data, k=5, algorithm=1)
+
+    assert dtc1[1] == pytest.approx(mi1[1], abs=pytest_abs)
