### 2025-11-19 1fc887c
KNN DTC
