### 2025-11-20 0f07c9f
KSG CMI
