+++ syntropy/knn/__init__.py
@@ -1,7 +1,7 @@
 from .shannon import (
     differential_entropy,
     mutual_information,
-    #conditional_mutual_information
+    conditional_mutual_information
 )
 
 from.multivariate_mi import (
@@ -14,7 +14,7 @@ from.multivariate_mi import (
 __all__ = [
     "differential_entropy",
     "mutual_information",
-    #"conditional_mutual_information",
+    "conditional_mutual_information",
      "total_correlation",
      "dual_total_correlation",
     # "s_information",
