### 2025-11-27 thosvarley/syntropy#3 @cvanelteren
@thosvarley ping me if you need help on engineering. I am happy to put some time in if you need support for the software engineering side.
