### 2025-12-20 thosvarley/syntropy#3 @cbueth
@thosvarley How are you doing, have you had time to go through the points noted?
