### 2026-01-11 thosvarley/syntropy#3 @thosvarley
@cbueth Sorry, the end of the semester and the holidays got away from me. I will try and get work done this month.
