+++ docs/source/conf.py
@@ -13,7 +13,7 @@ sys.path.insert(0, os.path.abspath('../..'))
 project = 'Syntropy'
 copyright = '2025, Thomas F. Varley'
 author = 'Thomas F. Varley'
-release = '1.0.0'
+release = '0.0.1'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
