+++ docs/source/api/syntropy.mixed.rst
@@ -1,0 +1,26 @@
+syntropy.mixed package
+====================
+
+.. automodule:: syntropy.mixed
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Submodules
+----------
+
+syntropy.mixed.shannon module
+---------------------------
+
+.. automodule:: syntropy.mixed.shannon
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+syntropy.mixed.utils module
+-------------------------
+
+.. automodule:: syntropy.mixed.utils
+   :members:
+   :show-inheritance:
+   :undoc-members:
