### 2026-01-16 bb7dff7
Mixed function docs
