+++ README.md
@@ -1,85 +1,177 @@
 # Syntropy
 
-Syntropy is a Python package for information-theoretic analysis of discrete and continuous random variables. 
+[![Python 3.8+](https://img.shields.io/badge/python-3.8+-blue.svg)](https://www.python.org/downloads/)
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+[![Documentation](https://readthedocs.org/projects/syntropy/badge/?version=latest)](https://syntropy.readthedocs.io/en/latest/)
 
-Wherever possible, syntropy provides discrete and Gaussian estimators for a given function (entropy, mutual information, total correlation, etc). 
+**Syntropy** is a Python library for multivariate information-theoretic analysis of discrete and continuous data. It provides efficient implementations of information measures ranging from basic quantities like entropy and mutual information to modern constructs like the partial information decomposition, O-information, and information rates for time series.
 
-#### Documentation 
-Read the documentation [here](https://syntropy.readthedocs.io/en/latest/).
+## Features
 
-#### Estimators 
+- **Multiple estimators**: Discrete, Gaussian, KNN (Kraskov), and neural (normalizing flow) estimators
+- **Pointwise measures**: Access local/pointwise values, not just expected values
+- **Higher-order information**: Total correlation, dual total correlation, O-information, S-information
+- **Information decomposition**: Partial information decomposition (PID) and partial entropy decomposition
+- **Time series**: Information rates and Lempel-Ziv complexity measures
+- **Consistent API**: Same interface across all estimator types
 
-Information theoretic analysis requires estimating entropies from different kinds of data - discrete and continuous. This is a non-trivial problem, especially in the case of real-valued data, as the underlying probability distributions can be non-parametric. The Syntropy package implements functions with three classes of estimators: discrete, continuous using Gaussian assumptions, and continuous estimators using normalizing flows.  
+## Installation
 
-##### Discrete 
+```bash
+git clone https://github.com/thosvarley/syntropy.git
+cd syntropy
+pip install .
+```
 
-The discrete estimators are built on discrete probability distributions and the measures computed directly:
+For development:
+```bash
+pip install -e ".[dev]"
+```
 
-$$H(X) = -\sum_{x\in\mathcal{X}}P_X(x)\log P_X(x)$$
+## Quick Start
 
-The distirbutions are represented in Python as dictionaries where the keys are tuples representing the joint-state of every element, and the values are the probabilities of that state. For example, an XOR distribution would look like:
+### Discrete Distributions
+
+Discrete estimators work with probability distributions represented as dictionaries:
 
 ```python
-xor_distribution = {
-    (0,0,0) : 1/4,
-    (0,1,1) : 1/4,
-    (1,0,1) : 1/4,
-    (1,1,0) : 1/4,
+from syntropy.discrete import mutual_information, o_information
+
+# XOR distribution: pure synergy
+xor = {
+    (0, 0, 0): 0.25,
+    (0, 1, 1): 0.25,
+    (1, 0, 1): 0.25,
+    (1, 1, 0): 0.25,
+}
+
+# Mutual information between inputs (0,1) and output (2)
+ptw, mi = mutual_information(idxs_x=(0, 1), idxs_y=(2,), joint_distribution=xor)
[…]
