+++ docs/source/theory.rst
@@ -1,0 +1,266 @@
+Information Theory Primer
+=========================
+
+This document provides a conceptual introduction to the core ideas of information theory, with particular attention to the challenges of estimation and higher-order information measures. For formal treatments, see Shannon's original work or Cover & Thomas's *Elements of Information Theory*. Much of the content here is based on the review article `"Information Theory for Complex Systems Scientists" <https://arxiv.org/abs/2304.12482>`_.
+
+Entropy: Quantifying Uncertainty
+--------------------------------
+
+Shannon entropy is the foundational quantity of information theory. It quantifies the uncertainty associated with a random variable—how "surprised" we should expect to be when we observe its outcome.
+
+The Intuition
+^^^^^^^^^^^^^
+
+There are two complementary ways to think about entropy:
+
+**Expected Surprise.** Rare events are surprising; common events are not. The "surprise" or "self-information" of observing an outcome :math:`x` with probability :math:`P(x)` is:
+
+.. math::
+
+   h(x) = -\log P(x)
+
+If an outcome is certain (:math:`P(x) = 1`), there is no surprise (:math:`h(x) = 0`). If an outcome is very rare, observing it is highly surprising. Entropy is simply the *expected* surprise across all possible outcomes:
+
+.. math::
+
+   H(X) = \mathbb{E}_X[h(x)] = -\sum_{x \in \mathcal{X}} P(x) \log P(x)
+
+**Minimum Questions.** Entropy can also be interpreted as the minimum average number of yes/no questions needed to identify the state of a random variable. A fair 6-sided die requires about 2.58 questions on average, while a biased die (where one face appears most of the time) requires fewer questions—you can usually guess the common outcome first.
+
+Key Properties
+^^^^^^^^^^^^^^
+
+- **Non-negativity:** :math:`H(X) \geq 0` (for discrete variables).
+- **Maximum at uniformity:** Entropy is maximized when all outcomes are equally probable.
+- **Additivity for independent variables:** If :math:`X_1` and :math:`X_2` are independent, then :math:`H(X_1, X_2) = H(X_1) + H(X_2)`.
+- **Subadditivity:** In general, :math:`H(X_1, X_2) \leq H(X_1) + H(X_2)`, with equality only when the variables are independent.
+
+Mutual Information: Quantifying Dependence
+------------------------------------------
+
+While entropy measures uncertainty about a single variable, mutual information measures the statistical dependence between two variables—how much knowing one reduces uncertainty about the other.
+
+Definition and Interpretations
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Mutual information can be expressed in several equivalent ways:
+
+.. math::
+
+   I(X; Y) &= H(X) - H(X|Y) \\
+           &= H(Y) - H(Y|X) \\
+           &= H(X) + H(Y) - H(X, Y)
+
+The first form says: mutual information is how much the entropy of :math:`X` decreases when we learn :math:`Y`. The third form shows it as the "overlap" between the individual entropies, minus the joint entropy.
+
+**Bayesian Interpretation.** Mutual information equals the Kullback-Leibler divergence between the true joint distribution and the product of marginals:
+
+.. math::
+
+   I(X; Y) = D_{KL}(P(X,Y) \| P(X) P(Y))
+
+This measures how much information we gain when updating from an assumption of independence to the actual joint distribution.
+
+Local Mutual Information
+^^^^^^^^^^^^^^^^^^^^^^^^
+
+The *pointwise* or *local* mutual information for specific observations :math:`x` and :math:`y` is:
+
+.. math::
+
+   i(x; y) = \log \frac{P(x, y)}{P(x) P(y)} = \log \frac{P(x|y)}{P(x)}
+
[…]
