### 2026-01-16 1485b3c
Theory primer
