### 2026-01-16 2ecffcf
Theory primer
