### 2026-01-16 ed82839
KNN-based mixed datatype estimator
