### 2026-01-18 thosvarley/syntropy#3 @thosvarley
@cbueth I've tried to address as many of your comments as I could. I've now got a CONTRIBUTING.md and CODE_OF_CONDUCT.md file. The pyproject.toml file is more complete, and I've put in numbered dependencies.

The issue of the unit tests is a bit tricky - I can add tests that ensure that you get the right types out, but for a lot of these, I'm not sure exactly what the best way to write ground-truth tests. I could just generate Gaussian data with fixed covariance structures, but due to things like bias/variance in the estimators, noise and finite size effects, I don't think I can get reliable comparisons to analytic ground truth the way I can with the Gaussian and discrete estimators.

I'm certianly open to any ideas in this vein, however.
