### 2026-01-18 thosvarley/syntropy#3 @cvanelteren
I think they key for this package would be to test for simple variables and that they produce the correct outcome. With unittests you can also check if it is within bounds -- consistency can be gained by just fixing the seed for more complex examples.
