### 2026-01-18 thosvarley/syntropy#3 @cbueth
Hey @thosvarley, thanks for the update! Just to makes sure I'll look at the right changes, they are the commits on the main branch of the last two days?

I agree with what @cvanelteren said, you already give the answer yourself 😁 In that case it would be nice to test with a gaussian and expecting the theoretical result within some bounds. We had the same problem for the `infomeasure` package. There we also added the gaussian case https://github.com/cbueth/infomeasure/blob/main/tests/estimators/entropy/test_gaussian_distribution.py, but also test against precomputed values for stability.
While you don't need to test everything, it would be great to have tests for the most central and common elements. If the output is noisy, you can always test with relative and absolute tolerances (see https://docs.pytest.org/en/stable/how-to/assert.html#assertraises)
