### 2026-01-18 thosvarley/syntropy#3 @thosvarley
Yes, everything is just getting pushed to main. At some point I'll split it into a dev/ and main/ branch, but for now it's all in one.

With regards to the name - I am kicking myself that "syntropy" was taken on PyPI in November. I actually emailed the owner of that package asking if they would release the name out of the goodness of their heart. Depending on what they say (if they reply at all), I'll cross that bridge when I get to it.
