### 2026-01-19 de7b799
Updating install instructions
