+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "syntropy"
+name = "syntropyx"
 version = "0.0.1"
 authors = [
   { name="Thomas F. Varley", email="tfvarley@uvm.edu" },
@@ -41,7 +41,7 @@ dependencies = [
 [project.urls]
 Homepage = "https://github.com/thosvarley/syntropy"
 Repository = "https://github.com/thosvarley/syntropy"
-Documentation = "https://syntropy.readthedocs.io"
+Documentation = "https://syntropyx.readthedocs.io"
 Issues = "https://github.com/thosvarley/syntropy/issues"
 
 [project.optional-dependencies]
@@ -49,6 +49,9 @@ dev = [
     "pytest>=7.0.0",
     "pytest-cov>=4.0.0",
 ]
+
+[tool.hatch.build.targets.wheel]
+packages = ["syntropy"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
