### 2026-01-19 f820d68
Gaussian copula estimators
