### 2026-01-19 4ff203a
Gaussian copula estimators
