### 2026-01-21 7b8a5d1
Discrete estimators
