+++ syntropy/discrete/estimators.py
@@ -1,0 +1,490 @@
+import numpy as np
+
+from scipy.special import digamma
+from numpy.typing import NDArray
+from typing import Any, Sequence, Callable
+
+from .shannon import shannon_entropy, mutual_information
+from .utils import get_marginal_distribution
+
+DiscreteDist = dict[tuple[Any, ...], float]
+
+
+def dirichlet_probabilities(
+    data: NDArray[Any],
+    prior: str | Sequence[float],
+    alphabet: Sequence[Any] | None = None,
+) -> DiscreteDist:
+    """
+    Estimate probabilities using Dirichlet-Multinomial Bayesian model.
+
+    Parameters
+    ----------
+    data : NDArray[Any]
+        2D array of shape (channels, samples) containing discrete observations.
+    prior : str | Sequence[float]
+        Either a named prior ('mle', 'laplace', 'jeffreys', 'perks') or a 
+        custom vector of pseudocounts with length matching the alphabet size.
+    alphabet : Sequence[Any] | None
+        Optional full alphabet of possible states. If None, uses only observed states.
+
+    Returns
+    -------
+    DiscreteDist
+        Dictionary mapping states to their estimated probabilities. Only includes
+        states with non-zero probability.
+
+    References
+    ----------
+    Perks, W. (1947). Some observations on inverse probability including a new indifference rule. Journal of the Institute of Actuaries, 73, 285-334.
+    
+    Hausser, J., & Strimmer, K. (2009). Entropy inference and the James-Stein estimator, with application to nonlinear gene association networks. 
+    Journal of Machine Learning Research, 10, 1469-1484. 
+    https://jmlr.org/papers/v10/hausser09a.html
+    """
+    assert len(data.shape) == 2, "The data must be two-dimensional."
+
+    # Extract observed symbols (columns are observations)
+    unq, counts = np.unique(data, axis=1, return_counts=True)
+    observed_states = [tuple(unq[:, i]) for i in range(unq.shape[1])]
+    observed_counts = dict(zip(observed_states, counts))
+
+    N = data.shape[1]
+
+    states: Sequence[Any]
+    counts_aligned: NDArray[np.integer]
+    # Determine full support and aligned counts
+    if alphabet is None:
+        states = observed_states
+        counts_aligned = np.array([observed_counts[s] for s in states])
+    else:
+        states = alphabet  # fixed, deterministic order
+        counts_aligned = np.array([observed_counts.get(s, 0) for s in states])
+
+    k: int = len(states)
+    alpha: float
+    alpha_vec: NDArray[np.floating]
+    if isinstance(prior, str):
+        assert prior in {"mle", "laplace", "jeffreys", "perks"}
+
+        if prior == "mle":
+            alpha = 0.0
+        elif prior == "laplace":
+            alpha = 1.0
+        elif prior == "jeffreys":
+            alpha = 0.5
+        elif prior == "perks":
+            alpha = 1.0 / k
+
+        alpha_vec = np.full(k, alpha)
+    else:
+        alpha_vec = np.array(prior, dtype=float)
+        if len(alpha_vec) != k:
+            raise ValueError(
+                f"Length of prior ({len(alpha_vec)}) must match number of states ({k})"
+            )
+
+    probs = (counts_aligned + alpha_vec) / (N + alpha_vec.sum())
+
+    return {states[i]: probs[i] for i in range(k) if probs[i] > 0}
+
+
+def plugin_probabilities(data: NDArray[Any]) -> DiscreteDist:
+    """
+    Maximum likelihood (plugin) probability estimator.
+
[…]
