### 2026-01-26 #9407 @cbueth
Hey @crvernon, thanks for the message. Yes, I am able to continue the review. @thosvarley got back to my first points last week (https://github.com/thosvarley/syntropy/issues/3#issuecomment-3764744679), since then I have not found the time but have this on my mind.
