### 2026-01-26 thosvarley/syntropy#3 @cbueth
As I'm a bit busy these days, have the changes regarding my first points been addressed in the pushed commits? I am just asking to know when it's the right time to do a second pass, looking over them or to wait a bit more until you ping me again.
