### 2026-01-26 thosvarley/syntropy#3 @thosvarley
I think the last thing is the problem of unit tests, which may take me a little while, as I am also v. busy these days.
