### 2026-01-26 thosvarley/syntropy#3 @cbueth
Ok, then I might look before as I can check functionalities independently of the tests anyways.
