### 2026-01-26 thosvarley/syntropy#3 @cbueth
I just took some tome today afternoon to look through the changes. Generally I think the package is taking shape, but there are still important requirements to be met. I try to summarise this shortly:

- [ ] A central and compact place describing which functionality is given (and for which estimating technique) - something globally one can grasp before reading though the quickstart guide or needing to dig through the API. While this is a bit more clear in the paper, this is missing from the documentation.
- [ ] Installation instructions in documentation
- [ ] Statement of need (in the doc)
- [ ] More tests
- [ ] Some points on the paper - see notes
- [ ] References to the community guidelines and code of conduct in the documentation pages.

Furthermore, there are some specific points on the documentation I have for you. What follows are my notes reviewing the package. I hope they are not too chaotic, and I could structure them somehow clearly. If there are any unclear points, please ask!

**Edit:** When reviewing I did not read the README.md (my bad!). The table you have https://github.com/thosvarley/syntropy?tab=readme-ov-file#available-measures and the list https://github.com/thosvarley/syntropy?tab=readme-ov-file#features with all the functionality is helpful. I was missing such compact functionality summary from the documentation or paper. So some of the following points might be due because I did not consider the `README.md`, but that might happen to a normal user, too :)

### Examples:

I was able to reproduce the examples from the documentation index on python `3.14`, like so:

```
>>> from syntropy.discrete import mutual_information
... from syntropy.discrete.distributions import XOR_DIST
...
... ptw, avg = mutual_information(
...      idxs_x = (0,1),
...      idxs_y = (2,),
...      joint_distribution=XOR_DIST
...      )
... print(f"I(X1,X2 ; Y) = {avg}") # Equal to 1 bit.
...
/Users/cbueth/projects/reviews/syntropy/syntropy/discrete/decompositions.py:212: SyntaxWarning: "\c" is an invalid escape sequence. Such sequences will not work in the future. Did you mean "\\c"? A raw string is also an option.
  :math:`h^{sx}_{\cap}(\\alpha) = -\\log_{2} P(a_{1} \\cup a_{2} \\cup \\ldots \\cup a_{k})`.
/Users/cbueth/projects/reviews/syntropy/syntropy/discrete/decompositions.py:514: SyntaxWarning: "\m" is an invalid escape sequence. Such sequences will not work in the future. Did you mean "\\m"? A raw string is also an option.
  i_{\\min}(\\alpha;t) &= \\min h(\\alpha_i) - \min h(\\alpha_i|t) \\\\
/Users/cbueth/projects/reviews/syntropy/syntropy/gaussian/temporal.py:61: SyntaxWarning: "\i" is an invalid escape sequence. Such sequences will not work in the future. Did you mean "\\i"? A raw string is also an option.
  H(X)=\\frac{1}{4\\pi} \int_{-\\pi}^{\\pi} \\log \\left( (2\\pi e)^N |S_X(\\omega)| \\right) \\, d\\omega
I(X1,X2 ; Y) = 1.0
```

I guess these warning should be fixed by using different docstrings `r"""` instead of `"""`. I was not able to check how this function (`hsx_discrete_redundancy`) looks in the documentation, as it seems like there is no documentation generated for it. You can consider using `sphinx-apidoc` (https://www.sphinx-doc.org/en/master/man/sphinx-apidoc.html). This can be used to generate full API documentation rst files. These then can be included to be compiled with `sphinx-build`. But the current approach with `autodoc` and only generating the API for the functions to be used by the user is also valid.

```
>>> import numpy as np
>>> from scipy.stats import multivariate_normal
... from syntropy.gaussian import mutual_information
...
... num_samples = 100_000
... cov = np.array([
...      [0.99999999, 0.24404644, 0.65847509],
...      [0.24404644, 0.99999985, 0.24163274],
...      [0.65847509, 0.24163274, 0.99999996]
...      ])
... data = multivariate_normal.rvs(
...      mean = [0,0,0],
...      cov = cov,
...      size=num_samples
...      ).T
... avg = mutual_information(
...      idxs_x=(0,1),
...      idxs_y=(2,), cov=np.cov(data, ddof=0.0)
...      )
... print(f"I(X;Y) = {avg:.3} nat")
...
I(X;Y) = 0.289 nat
>>> # Example with Kraskov estimator
... from syntropy.knn import mutual_information
...
... num_samples = 10_000
... rand = np.random.randn(num_samples)
... data = np.vstack((rand, rand))
... data[1,:] = (data[0,:]**2) + np.random.randn(num_samples)
...
... ptw, avg = mutual_information(
...      idxs_x=(0,),
...      idxs_y=(1,),
...      data=data,
...      k = 5,
...      algorithm = 1 # Whether to use the KSG1 or KSG2 MI estimator.
... )
...
... print(f"I(X;Y) = {avg:.3} nat")
...
I(X;Y) = 0.435 nat
```

I'd like to suggest that you indluce the expected outputs in the example, so a user can cross validate it works on their side, and for users who only read the documentation, wito

### Guide

The Information Theory Primer seems to be useful for a new user, but also to be sure about the nomenclature used in this package.

#### Quickstart Guide

**Discrete estimators:**

- I think the API you defined is clever, beginner-friendly (if they have some info theory background) and useful, returning the local and global values. They work functionally (opposed to object oriented estimators).
- The output format is useful, because you always have the indices with the local values. While this is surely useful, this also increases the data the computer must handle.
- "The basic form of the discrete estimators is a Python dictionary, which represents the joint probability distribution over some set of elements." The discrete estimators are not the dictionary of the probability distribution. Discrete estimators work with discrete distributions, but they are not the input data themselves.
- I think this subsection can still be improved to be more clear for a first time user.

**Gaussian estimators:**

- "in my cases, expected values can be computed from the covariance matrix directly" This sounds like the separation of local and global values is especially useful for you/your data. If this is generally the case, please specify it in this sentence.
- This example does not show what the expected out put is. Please deliver the expected output for all examples.

**KNN:**

- I was able to swap X and Y and got the same output 👍

**Neural estimators:**

- When starting, I am getting `/torch/nn/modules/linear.py:124: UserWarning: Initializing zero-element tensors is a no-op
  init.kaiming_uniform_(self.weight, a=math.sqrt(5))`
- The second cell fails for me with: `syntropy/neural/multivariate_mi.py`, line 53, in total_correlation `TypeError: differential_entropy() got an unexpected keyword argument 'context'`.

**Mixed estimators:**

- Example works (if you remove the last dot)

Just as a suggestion, don't feel obliged: It is possible to include jupyter notebooks in Sphinx pages where you can write code cells which then show the output pretty. You can find information about it here https://executablebooks.org/en/latest/ and how to import it to Sphinx is here https://myst-nb.readthedocs.io/en/latest/quickstart.html. It also supports adding code cells to `.rst` files, as you have them now.

It seems like there are more functionalities that have not been covered in the guide page. I'd suggest you to add an overview somewhere what features there are, what the bigger picture there is and intent of the package - independent of the paper. If you have future functionality you want to add, you can also put them on a roadmap.

The langing page has a description that introduces well what this package is about, what I am missing are instructions or examples on how to calculate "Kullback-Leibler divergence, to modern constsructs like the partial information decomposition, higher-order measures (e.g. the O-information)" There is an O-information example, but only for the neural case. Is this the only available one? (I have not worked with it much, but rather with (conditional) mutual/ interaction information, TE and so on)

What is also still needed are clear installation instructions and a statement of need in the documentation. Both are required by the JOSS guidelines.

### Documentation Formatting:

- In the documentation you are including a non-existent module. Does it exist uncommited? Otherwise you might want to delete it. You can see the empty heading here: https://syntropy.readthedocs.io/en/latest/api/syntropy.mixed.html#syntropy-mixed-utils-module And it causes a warning when building the docs https://app.readthedocs.org/projects/syntropy/builds/31095042/.

  https://github.com/thosvarley/syntropy/blob/afa480c2cd64807cc5f1b02eecf327ae605151e0/docs/source/api/syntropy.mixed.rst#L20-L26

- Furthermore ther are several warnings when you build the documentation (https://app.readthedocs.org/projects/syntropy/builds/31095042/), please resolve these. One appears because the API documentation is included twice. The Sphinx documentation system has this requirement to have a clear document tree, when you include documents twice. You have:

  https://github.com/thosvarley/syntropy/blob/0bce05d8f14f5817ce921774ff2812ed9d58896f/docs/source/index.rst#L80-L88

  And:

  https://github.com/thosvarley/syntropy/blob/0bce05d8f14f5817ce921774ff2812ed9d58896f/docs/source/api/syntropy.rst#L9-L20

  You can see this in the log here `/docs/source/api/syntropy.discrete.rst: document is referenced in multiple toctrees: ['api/syntropy', 'index'], selecting: index <- api/syntropy.discrete
/`. The most common structure is to let the `docs/source/index.rst` toctree include the `docs/source/api/syntropy.rst` file. As of now I was not able to navigate to the `docs/source/api/syntropy.rst` page itself. Is it possible somehow? As you use `automodule` you also probably do not need `docs/source/api/syntropy.mixed.rst` in the first place.

  Just to name another warning, e.g. `/docs/source/api/modules.rst: WARNING: document isn't included in any toctree [toc.not_included]`. Furthermore, a changelog is being referenced but does not exist

  Before you mentioned you're not a software engineer. This is not a problem, just read the warnings and see where they point you. Let me know if this is incomplete. Before a publication you might want to have this rounded up.

### Repo & Packaging:

I think the new commits were able to improve the package, especially the top-level `__init__.py` which imports the submodules, the changed package version and package information.
For the adapted code in `syntropy/knn/multivariate_mi.py` it is good you referenced the implementation base you used from JIDT.

- You have added some items to the `.gitignore`, but the committed folders and files still exist, so you need to explicetly remove them. I can also see that the files in `build/` have been updated after you added them to the `.gitignore`, in 0bce05d8. This is unually not included, as the only relevant files are the sources of which the documentation can be built. Furthermore, `syntropy/mixed/shannon.py` is also in the `.gitignore` file, you surely do not want to ignore the source code.
- The `pyproject.toml` specifies python versions 3.8-3.12. The current python version is 3.14 and support for 3.10 is being dropped for many actively maintained packages as being end-of-life this year. Please consider supporting the current python version. See https://devguide.python.org/versions/.
- Nice you have added `coverage` 👍
- ~~Is the package naming is still open? Let me know if this issue is settled. Currently I see some references to `syntropy` and some to `syntropyx`.~~ At the installttion instruction, you can make the note more clear there is a difference in naming with such note https://github.com/orgs/community/discussions/16925.
- While the code overall looks to be in good standard, there is still some formatting potential. I'd suggest you to use some kind of linting and code formatting. An example would be `ruff` https://docs.astral.sh/ruff/. With it you can easily format the code, find some common mistakes and solve some of the automatically. One example would be unused imports (e.g. the unused `scipy` https://github.com/thosvarley/syntropy/blob/a7fdb2dce32feb297cf4851acbcaac54f39b47be/syntropy/neural/multivariate_mi.py#L1 can be automatically fixed). The next step would be type checking, but I think for this project it would overshoot the target.
- The package description claims the implementations are _efficient_. In what way and can you substantiate that somehow? A comparison to other packages is always nice, but I also see that this might not be the selling point, as you have a variety of functionality rather than, let say, GPU support. Edit: Is this referencing this https://github.com/thosvarley/syntropy?tab=readme-ov-file#optimizations-and-utilities ?

- `README.md`: I have this problem with the first example:

```
Python 3.14.2 (main, Jan 14 2026, 23:37:46) [Clang 21.1.4 ] on darwin
>>> from syntropy.discrete import mutual_information, o_information
...
... # XOR distribution: pure synergy
... xor = {
...     (0, 0, 0): 0.25,
...     (0, 1, 1): 0.25,
...     (1, 0, 1): 0.25,
...     (1, 1, 0): 0.25,
... }
...
... # Mutual information between inputs (0,1) and output (2)
... ptw, mi = mutual_information(idxs_x=(0, 1), idxs_y=(2,), joint_distribution=xor)
... print(f"I(X0,X1 ; X2) = {mi:.3f} bits")  # 1.0 bit
...
... # O-information (negative = synergy-dominated)
... ptw, omega = o_information(idxs=(0, 1, 2), joint_distribution=xor)
... print(f"Omega = {omega:.3f} bits")  # -1.0 bit
...
I(X0,X1 ; X2) = 1.000 bits
Traceback (most recent call last):
  File "<python-input-0>", line 16, in <module>
    ptw, omega = o_information(idxs=(0, 1, 2), joint_distribution=xor)
                 ~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
TypeError: o_information() got an unexpected keyword argument 'idxs'
```

All other examples work :) It would be nice to also show the result, so one know what to expect.

## Paper Comments

- The paper includes the statement of need, and has substance 👍
- The citations need correct formatting. When citing multiple separate them by semicolon, using `[@goodwell_its_2020; @goodwell_debatesdoes_2020]`, not `[@goodwell_its_2020][@goodwell_debatesdoes_2020]`,  (see pandoc manual https://pandoc.org/MANUAL.html#extension-citations).
- Maybe this package might be interesting https://brainets.github.io/hoi/index.html
- In the paper you mention the package is written in pure python, while you mention `numpy`, `scipy` and `networkx` are used, `torch` and `nflows` are missing. While I think it's still a small number of packages (nice!), the current wording "pure python" can be misleading.
- About what packages there are, we also looked into this not too long ago, see the table of our paper https://www.nature.com/articles/s41598-025-14053-5/tables/1. Maybe this might be helpful.
- Suggestion: Add a short example into the paper, maybe with some real or gaussian data and visualisation.
