### 2026-01-26 #9407 @cbueth
Today I was able to do my second round of review, this time also testing the software. I was able to install it and look more into the code. Some improvements from my first short round were already implemented. While I could cross off further points off of my checklist, there are still some missing that I reminded the author about, among further improvements. I do have the impression that the author has a solid background in the field and there is a scholarly effort in the software. While the paper seems to have everything fundamentally important in it, the documentation and testing are still lacking, but are on a promising path.
