### 2026-01-29 6319faf
Debug neural estimators
