+++ syntropy/neural/multivariate_mi.py
@@ -1,19 +1,18 @@
-import scipy.stats as stats
 import torch
 from .shannon import differential_entropy
 
 
+# %%
 def total_correlation(
     idxs: tuple[int],
     data: torch.Tensor,
-    context: None | tuple[int] = None,
     data_test: None | torch.Tensor = None,
     flow_kwargs: dict = None,
     train_kwargs: dict = None,
     verbose: bool = False,
-) -> tuple[float, float]:
+) -> tuple[torch.Tensor, float]:
     """
-    Computes the total correlation of the data using normalizing flow estimators.  
+    Computes the total correlation of the data using normalizing flow estimators.
 
 
     Parameters
@@ -22,8 +21,6 @@ def total_correlation(
         The tuple of indices the differential entropy is computed for.
     data : torch.Tensor
         The training data, in samples x features format.
-    context : None | tuple[int]
-        If not None, the indices of the conditioning variables.
     data_test : None | torch.Tensor
         If not None, the testing data in samples x features format.
     flow_kwargs : dict
@@ -36,6 +33,8 @@ def total_correlation(
 
     Returns
     -------
+    torch.Tensor
+
     float
 
 
@@ -43,29 +42,24 @@ def total_correlation(
     flow_kwargs = flow_kwargs or {}
     train_kwargs = train_kwargs or {}
 
-    if context is None:
-        context_arg = None
-    else:
-        context_arg = context
+    lookup: dict[tuple[int, ...], torch.Tensor] = {}
 
-    lookup: dict[tuple[int,...], float] = {}
-
-    h_idxs, _ = differential_entropy(
+    h: torch.Tensor
+    h, _, _ = differential_entropy(
         idxs=idxs,
         data=data,
-        context=context_arg,
         data_test=data_test,
         verbose=verbose,
         train_kwargs=train_kwargs,
         flow_kwargs=flow_kwargs,
     )
-    lookup[idxs] = h_idxs
+    lookup[idxs] = h
 
+    h_i: torch.Tensor
     for i in idxs:
-        h_i, _ = differential_entropy(
+        h_i, _, _ = differential_entropy(
             idxs=(i,),
             data=data,
-            context=context_arg,
             data_test=data_test,
             verbose=verbose,
             flow_kwargs=flow_kwargs,
@@ -74,28 +68,27 @@ def total_correlation(
 
         lookup[(i,)] = h_i
 
-    tc: float = 0.0
+    tc: torch.Tensor = torch.zeros(data.shape[0])
 
     for key in lookup.keys():
         if len(key) == 1:
-            tc += lookup[key][0]
+            tc += lookup[key]
         else:
-            tc -= lookup[key][0]
+            tc -= lookup[key]
 
-    return tc
+    return tc, tc.mean()
 
 
 def higher_order_information(
     idxs: tuple[int],
     data: torch.Tensor,
-    context: None | tuple[int] = None,
     data_test: None | torch.Tensor = None,
     flow_kwargs: dict = None,
     train_kwargs: dict = None,
     verbose: bool = False,
 ) -> dict[str, float]:
     """
-    Computes the O-information, S-information, total correlation, and dual total correlation for the data. 
+    Computes the O-information, S-information, total correlation, and dual total correlation for the data.
     Computing them all as a set is more efficient than computing each one independently.
 
     Parameters
@@ -114,44 +107,35 @@ def higher_order_information(
         Arguments for the utils.train_flow function.
     verbose : bool
[…]
