### 2026-01-30 1326ec7
KNN tests + improving function argument defaults
