+++ syntropy/knn/multivariate_mi.py
@@ -2,12 +2,14 @@ import numpy as np
 from numpy.typing import NDArray
 from scipy.special import digamma
 from scipy.spatial import cKDTree
-from .utils import check_idxs, build_tree_and_get_distances, get_counts_from_tree
+from .utils import build_tree_and_get_distances, get_counts_from_tree
+from ..utils import check_idxs
+
 
 def total_correlation(
     data: NDArray[np.floating],
     k: int,
-    idxs: tuple[int, ...] = (-1,),
+    idxs: tuple[int, ...] | None = None,
     algorithm: int = 1,
 ) -> tuple[NDArray[np.floating], float]:
     """
@@ -48,7 +50,7 @@ def total_correlation(
 
 
 def total_correlation_1(
-    data: NDArray[np.floating], k: int, idxs: tuple[int, ...] = (-1,)
+    data: NDArray[np.floating], k: int, idxs: tuple[int, ...] | None = None
 ) -> tuple[NDArray[np.floating], float]:
     r"""
     Computes the Kraskov, Stogbauer, Grassberger estimate of the total correlation using the first algorithm presented Kraskov et. al (2004)
@@ -85,9 +87,9 @@ def total_correlation_1(
     https://doi.org/10.1147/rd.41.0066
 
     """
-    idxs: tuple[int, ...] = check_idxs(idxs, data.shape[0])
+    idxs_: tuple[int, ...] = check_idxs(idxs, data)
 
-    m: int = len(idxs)
+    m: int = len(idxs_)
     N: int = data.shape[1]
 
     psi_k: float = digamma(k)
@@ -96,12 +98,14 @@ def total_correlation_1(
     ptw: NDArray[np.floating] = np.full((1, N), psi_k + (m - 1) * psi_N)
 
     distances: NDArray[np.floating]
-    _, distances, _ = build_tree_and_get_distances(data=data[idxs, :], k=k)
+    _, distances, _ = build_tree_and_get_distances(data=data[idxs_, :], k=k)
 
-    for idx in idxs:
+    for idx in idxs_:
         tree, _, _ = build_tree_and_get_distances(data[(idx,), :], k=k)
         counts: NDArray[np.integer] = get_counts_from_tree(
-            tree, data[(idx,), :], distances[:, -1],
+            tree,
+            data[(idx,), :],
+            distances[:, -1],
         )
 
         ptw[0, :] -= digamma(counts + 1)
@@ -110,7 +114,7 @@ def total_correlation_1(
 
 
 def total_correlation_2(
-    data: NDArray[np.floating], k: int, idxs: tuple[int, ...] = (-1,)
+    data: NDArray[np.floating], k: int, idxs: tuple[int, ...] | None = None
 ) -> tuple[NDArray[np.floating], float]:
     r"""
     Computes the Kraskov, Stogbauer, Grassberger estimate of the total correlation using the second algorithm presented in Kraskov et. al., (2004).
@@ -147,9 +151,9 @@ def total_correlation_2(
     https://doi.org/10.1147/rd.41.0066
 
     """
-    idxs = check_idxs(idxs, data.shape[0])
+    idxs_: tuple[int, ...] = check_idxs(idxs, data)
 
-    m: int = len(idxs)
+    m: int = len(idxs_)
     N: int = data.shape[1]
 
     psi_k: float = digamma(k)
@@ -161,10 +165,10 @@ def total_correlation_2(
 
     distances: NDArray[np.floating]
     indices: NDArray[np.integer]
-    _, distances, indices = build_tree_and_get_distances(data[idxs, :], k=k)
[…]
