+++ syntropy/utils/__init__.py
@@ -1,0 +1,3 @@
+from .utils import (
+        check_idxs
+        )
