### 2026-01-31 759bd8d
Gaussian co-information
