+++ syntropy/utils/utils.py
@@ -1,5 +1,7 @@
 import numpy as np 
 from numpy.typing import NDArray
+from typing import Iterable
+import itertools as it 
 
 def check_idxs(idxs: tuple[int, ...] | None, data: NDArray[np.floating]) -> tuple[int, ...]:
     
@@ -9,3 +11,15 @@ def check_idxs(idxs: tuple[int, ...] | None, data: NDArray[np.floating]) -> tupl
         idxs_ = idxs 
 
     return idxs_
+
+def make_powerset(iterable: Iterable):
+    """
+    Computes the powerset of a collection of elements.
+
+    .. math::
+        \\mathcal{P}(\\{X_1,X_2,X_3\\}) \\to (\\{\\}, \\{X_1\\}, \\{X_2\\}, \\{X_3\\}, \\{X_1,X_2\\}, \\{X_1,X_3\\}, \\{X_1,X_2,X_3\\} )
+
+    """
+    xs: list = list(iterable)
+    # note we return an iterator rather than a list
+    return it.chain.from_iterable(it.combinations(xs, n) for n in range(len(xs) + 1))
