+++ README.md
@@ -156,6 +156,7 @@ print(f"I(discrete ; continuous) = {mi:.3f} nats")
 | Generalized Info. Decomp. | x | x | | |
 | Information Rates | x | x | | |
 | Connected Information | x | | | |
+| $\alpha$-synergy Decomp. | x | | | |
 
 ### Optimizations and Utilities
 
