### 2026-01-31 fff814c
Speed up differential entropy estimator
