+++ syntropy/gaussian/shannon.py
@@ -4,10 +4,10 @@ from numpy.typing import NDArray
 from .utils import check_cov
 from ..utils import check_idxs
 
-H_SINGLE: float = np.log(np.sqrt(2.0 * np.pi * np.e))
 LN_TWO_PI_E: float = np.log(2.0 * np.pi * np.e)
 TWO_PI: float = 2.0 * np.pi
 SQRT_TWO_PI: float = np.sqrt(TWO_PI)
+
 
 def differential_entropy(
     cov: NDArray[np.floating], idxs: tuple[int, ...] | None = None
@@ -15,7 +15,7 @@ def differential_entropy(
     """
     Computes the expected differential entropy of a multivariate
     distribution parameterized by a covariance matrix using a
-    Gaussian estimator.
+    Gaussian estimator via direct computation.
 
     The differential entropy is given by:
 
@@ -36,28 +36,30 @@ def differential_entropy(
     cov : NDArray[np.floating]
         The covariance matrix that defines the distribution.
     idxs : tuple[int, ...], optional
-        The specific subset of variables to compute the total correlation of.
-        Defaults to computing the TC of the entire covariance matrix.
+        The specific subset of variables to compute the entropy of.
+        Defaults to using the entire covariance matrix.
 
     Returns
     -------
     float
 
     """
+    idxs_: tuple[int, ...] = check_idxs(idxs, cov)
 
-    if idxs is None:
-        return stats.multivariate_normal(cov=cov, allow_singular=True).entropy()
-    else:
-        if len(idxs) == 1:
-            return H_SINGLE
-        else:
-            return stats.multivariate_normal(
-                cov=cov[np.ix_(idxs, idxs)], allow_singular=True
-            ).entropy()
+    cov_subset: NDArray[np.floating] = cov[np.ix_(idxs_, idxs_)]
+    k: int = len(idxs_)
+
+    # Use slogdet for numerical stability
+    sign, logdet = np.linalg.slogdet(cov_subset)
+
+    # H(X) = (k/2) * log(2πe) + (1/2) * log|Σ|
+    return 0.5 * (k * LN_TWO_PI_E + logdet)
 
 
 def local_differential_entropy(
-    data: NDArray[np.floating], cov: NDArray[np.floating] | None = None
+    data: NDArray[np.floating],
+    cov: NDArray[np.floating] | None = None,
+    idxs: tuple[int, ...] | None = None,
 ) -> NDArray[np.floating]:
     """
     Computes the framewise differential entropy for a set of variables.
@@ -77,6 +79,9 @@ def local_differential_entropy(
     cov : NDArray[np.floating], optional
         The covariance matrix that defines the distribution.
         If none is provided, it is computed from the data object.
+    idxs : tuple[int, ...], optional
+        The specific subset of variables to compute the entropy of.
+        Defaults to using all channels.
 
     Returns
     -------
@@ -84,18 +89,24 @@ def local_differential_entropy(
         The series of pointwise entropies.
 
     """
-    N: int = data.shape[0]
 
-    cov_: NDArray[np.floating] = check_cov(cov, data)
-
-    if N == 1:
+    if data.ndim == 1:
+        assert (idxs is None) or (idxs == (0,)), f"There's an indexing mismatch!"
         return -stats.norm.logpdf(
-            x=data, loc=data.mean(), scale=data.std()
-        )
+                x = data, 
+                loc = data.mean(),
+                scale = data.std()
+                )
     else:
+        idxs_: tuple[int, ...] = check_idxs(idxs, data)
[…]
