### 2026-02-01 4f36e99
Refactoring local TC
