### 2026-02-02 882351f
Gaussian I_dep decomposition
