+++ syntropy/gaussian/decompositions.py
@@ -1,8 +1,12 @@
 import numpy as np
 import networkx as nx
-from .shannon import local_differential_entropy
+from .shannon import (
+    local_differential_entropy,
+    mutual_information,
+    conditional_mutual_information,
+)
 
-from .utils import check_cov 
+from .utils import check_cov, covariance_to_correlation
 from ..utils import make_powerset
 from ..lattices import LATTICE_2, LATTICE_3, LATTICE_4
 from numpy.typing import NDArray
@@ -295,7 +299,9 @@ def mobius_inversion(
                     ]
                 )
 
-    ptw: dict[Atom, NDArray[np.floating]] = {node: lattice.nodes[node][partial_str] for node in lattice.nodes}
+    ptw: dict[Atom, NDArray[np.floating]] = {
+        node: lattice.nodes[node][partial_str] for node in lattice.nodes
+    }
     avg: dict[Atom, float] = {
         node: lattice.nodes[node][partial_str].mean() for node in lattice.nodes
     }
@@ -529,3 +535,132 @@ def representational_complexity(
 
 
 # %%
+
+def idep_partial_information_decomposition(
+    inputs: tuple[int, ...],
+    target: tuple[int, ...],
+    cov: NDArray[np.floating],
+) -> dict[str, float]:
+    """
+    Computes the I_dep partial information decomposition for Gaussian systems
+    using the dependency constraint method from Kay & Ince (2018).
+    
+    Currently only supports 2 predictors (univariate or multivariate).
+
+    Adapted from: https://github.com/robince/partial-info-decomp/blob/master/calc_pi_Idep_mvn.m
+    
+    Parameters
+    ----------
+    inputs : tuple[int, ...]
+        The indices of the two predictor variables/sets.
+        Must have length 2.
+    target : tuple[int, ...]
+        The indices of the target variable(s).
+    data : NDArray[np.floating] | None
+        The data in channels x samples format. Optional if cov provided.
+    cov : NDArray[np.floating] | None
+        The covariance matrix. If None, computed from data.
+        
+    Returns
+    -------
+    dict[str, float]
+        Dictionary with keys: 'unq0', 'unq1', 'red', 'syn'
+        
+    References
+    ----------
+    Kay, J. W., & Ince, R. A. A. (2018). 
+    Exact Partial Information Decompositions for Gaussian Systems 
+    Based on Dependency Constraints. Entropy, 20(4), 240.
+    https://doi.org/10.3390/e20040240
+    """
+    
+    assert len(inputs) == 2, "I_dep currently only supports 2 predictors"
+    
+    # Extract indices
+    input_0 = inputs[0]
+    input_1 = inputs[1]
+
+    n0, n1, nt = len(input_0), len(input_1), len(target)
+    
+    # Extract block covariances
+    C_00 = cov[np.ix_(input_0, input_0)]
+    C_11 = cov[np.ix_(input_1, input_1)]
+    C_tt = cov[np.ix_(target, target)]
+    C_01 = cov[np.ix_(input_0, input_1)]
+    C_0t = cov[np.ix_(input_0, target)]
+    C_1t = cov[np.ix_(input_1, target)]
+    
[…]

[…]
