### 2026-02-02 c53f17a
Typing on I_dep
