+++ syntropy/gaussian/scratch.py
@@ -1,166 +1,0 @@
-def idep_partial_information_decomposition(
-    inputs: tuple[int, ...],
-    target: tuple[int, ...],
-    data: NDArray[np.floating] | None = None,
-    cov: NDArray[np.floating] | None = None,
-) -> dict[str, float]:
-    """
-    Computes the I_dep partial information decomposition for Gaussian systems
-    using the dependency constraint method from Kay & Ince (2018).
-    
-    Currently only supports 2 predictors (univariate or multivariate).
-    
-    Parameters
-    ----------
-    inputs : tuple[int, ...]
-        The indices of the two predictor variables/sets.
-        Must have length 2.
-    target : tuple[int, ...]
-        The indices of the target variable(s).
-    data : NDArray[np.floating] | None
-        The data in channels x samples format. Optional if cov provided.
-    cov : NDArray[np.floating] | None
-        The covariance matrix. If None, computed from data.
-        
-    Returns
-    -------
-    dict[str, float]
-        Dictionary with keys: 'unq0', 'unq1', 'red', 'syn'
-        
-    References
-    ----------
-    Kay, J. W., & Ince, R. A. A. (2018). 
-    Exact Partial Information Decompositions for Gaussian Systems 
-    Based on Dependency Constraints. Entropy, 20(4), 240.
-    https://doi.org/10.3390/e20040240
-    """
-    
-    assert len(inputs) == 2, "I_dep currently only supports 2 predictors"
-    assert cov is not None or data is not None, "Must provide either data or covariance matrix"
-    
-    # Get covariance matrix
-    if cov is None:
-        cov = np.cov(data, ddof=0)
-    
-    # Extract indices
-    X0_idx = inputs[0] if isinstance(inputs[0], (list, tuple)) else (inputs[0],)
-    X1_idx = inputs[1] if isinstance(inputs[1], (list, tuple)) else (inputs[1],)
-    Y_idx = target if isinstance(target, (list, tuple)) else (target,)
-    
-    # Standardize to correlation matrix (Kay & Ince assume identity marginals)
-    joint_idx = tuple(X0_idx) + tuple(X1_idx) + tuple(Y_idx)
-    joint_cov = cov[np.ix_(joint_idx, joint_idx)]
-    corr = covariance_to_correlation(joint_cov)
-    
-    # Extract P, Q, R matrices (cross-correlations)
-    n0, n1, n2 = len(X0_idx), len(X1_idx), len(Y_idx)
-    
-    P = corr[:n0, n0:n0+n1]          # X0-X1 correlations
-    Q = corr[:n0, n0+n1:]            # X0-Y correlations  
-    R = corr[n0:n0+n1, n0+n1:]       # X1-Y correlations
-    
-    # Compute basic mutual informations using your functions
-    # (on standardized correlation matrix)
-    I_X0_Y = mutual_information(tuple(range(n0)), tuple(range(n0, n0+n2)), 
-                                 corr[:n0+n2, :n0+n2])
-    I_X1_Y = mutual_information(tuple(range(n0, n0+n1)), tuple(range(n0+n1, n0+n1+n2)),
-                                 corr[n0:, n0:])
-    I_X0X1_Y = mutual_information(tuple(range(n0+n1)), tuple(range(n0+n1, n0+n1+n2)), 
-                                   corr)
-    
[…]
