### 2026-02-24 aa1bff2
major update - PhiD, refactor antichain lattice, etc.
