+++ tests/test_gaussian.py
@@ -9,6 +9,7 @@ import syntropy.gaussian.shannon as shannon
 import syntropy.gaussian.multivariate_mi as mi
 from syntropy.gaussian.decompositions import (
     partial_information_decomposition as pid,
+    integrated_information_decomposition as phiid,
     partial_entropy_decomposition as ped,
     generalized_information_decomposition as gid,
     idep_partial_information_decomposition as idep,
@@ -28,7 +29,6 @@ cov = np.cov(data, ddof=0.0)
 # Due to natural instability in Scipy's matrix algebra, we need a slightly
 # more relaxed tolerance for our unit tests. 1 part in 1,000,000 is probably ok.
 pytest_abs = 1e-6
-
 
 def test_differential_entropy():
     h_all = shannon.local_differential_entropy(data).mean()
@@ -134,7 +134,7 @@ def test_pid():
 def test_ped():
     idxs = (0, 1, 2, 3)
 
-    ptw, avg = ped(idxs, data, cov)
+    ptw, avg = ped(data, idxs, cov)
     h = shannon.differential_entropy(cov, idxs)
 
     assert sum(avg.values()) == pytest.approx(h, abs=pytest_abs)
@@ -145,7 +145,7 @@ def test_gid():
     cov_prior = np.eye(len(idxs))
     cov_posterior = cov[np.ix_(idxs, idxs)]
 
-    ptw, avg = gid(idxs, data, cov_posterior, cov_prior)
+    ptw, avg = gid(idxs, data[idxs,:], cov_posterior, cov_prior)
 
     dkl = shannon.kullback_leibler_divergence(cov_posterior, cov_prior)
     assert sum(avg.values()) == pytest.approx(dkl, abs=pytest_abs)
@@ -157,45 +157,8 @@ def test_gid():
     assert ldkl.mean() == pytest.approx(sum(avg.values()), abs=pytest_abs)
 
 
-def test_gaussian_rate():
-    T = 5_000_000
-
-    noise = np.random.randn(2, T)
-
-    _, mir = mutual_information_rate((0,), (1,), noise, nperseg=2**13)
-
-    _, h1 = differential_entropy_rate((0,), noise, nperseg=2**13)
-    _, h2 = differential_entropy_rate((1,), noise, nperseg=2**13)
-    _, h12 = differential_entropy_rate((0, 1), noise, nperseg=2**13)
-
-    assert mir == pytest.approx(h1 + h2 - h12, abs=pytest_abs)
-
-    _, tcr = total_correlation_rate((0, 1), noise, nperseg=2**13)
-
-    assert mir == pytest.approx(tcr, abs=pytest_abs)
-
-    # for white noise with 0 mean and unit variance,
-    # the analytic entropy rate should be (1/2)*log(2*pi*e)
-    analytic = (1 / 2) * np.log(2.0 * np.pi * np.e)
-
-    assert analytic == pytest.approx(h1, abs=1e-3)
-
-    # %% A VAR(1) model with fixed noise and coefficient.
-
-    X = np.zeros((1, T))
-    eps = 0.01
-    noise = np.random.randn(T) * eps
-    a = 0.2
-
-    for t in range(1, T):
-        X[0, t] = X[0, t - 1] * a + noise[t - 1]
-
-    _, hX = differential_entropy_rate((0,), X, nperseg=2**14)
-
-    analytic = 0.5 * (np.log(2 * np.pi * np.e * (eps**2))) - (0.5 * np.log(1 - (a**2)))
-
-    assert np.isclose(hX, analytic, rtol=1e-2)
-
+def test_phiid():
+    return None
 
 def test_oinfo_rate():
     T = 5_000_000
[…]
