+++ tests/test_discrete.py
@@ -1,167 +1,323 @@
-import pytest 
+import pytest
 
 from syntropy.discrete.optimization import constrained_maximum_entropy_distributions
-from syntropy.discrete.distributions import (GIANT_BIT, XOR_DIST, SUMMED_DICE, 
-                                          RANDOM_DIST_2, RANDOM_DIST_3, RANDOM_DIST_4)
-from syntropy.discrete import multivariate_mi as mi 
+from syntropy.discrete.distributions import (
+    GIANT_BIT,
+    XOR_DIST,
+    SUMMED_DICE,
+    RANDOM_DIST_2,
+    RANDOM_DIST_3,
+    RANDOM_DIST_4,
+    RDNERR_DIST,
+)
+from syntropy.discrete import multivariate_mi as mi
 from syntropy.discrete import shannon
-from syntropy.discrete.decompositions import (partial_entropy_decomposition as ped, 
-                                           partial_information_decomposition as pid, 
-                                           generalized_information_decomposition as gid)
+from syntropy.discrete.decompositions import (
+    partial_entropy_decomposition as ped,
+    partial_information_decomposition as pid,
+    generalized_information_decomposition as gid,
+    integrated_information_decomposition as phiid,
+)
+
+pytest_abs = 1e-6
+
+# %%
 
 
 def test_total_correlation():
-    
     assert mi.total_correlation(GIANT_BIT)[1] == pytest.approx(2.0)
     assert mi.total_correlation(XOR_DIST)[1] == pytest.approx(1.0)
-    
+
     ptw, avg = mi.total_correlation(SUMMED_DICE)
-    
+
     assert avg == pytest.approx(3.2744019192887714)
-    assert sum(SUMMED_DICE[key]*ptw[key] for key in ptw.keys()) == pytest.approx(avg)
+    assert sum(SUMMED_DICE[key] * ptw[key] for key in ptw.keys()) == pytest.approx(avg)
 
 
 def test_dual_total_correlation():
-    
     assert mi.dual_total_correlation(GIANT_BIT)[1] == pytest.approx(1.0)
     assert mi.dual_total_correlation(XOR_DIST)[1] == pytest.approx(2.0)
-    
+
     ptw, avg = mi.dual_total_correlation(SUMMED_DICE)
-    
+
     assert avg == pytest.approx(5.169925001442313)
-    assert sum(SUMMED_DICE[key]*ptw[key] for key in ptw.keys()) == pytest.approx(avg)
+    assert sum(SUMMED_DICE[key] * ptw[key] for key in ptw.keys()) == pytest.approx(avg)
 
 
 def test_s_information():
-    
     assert mi.s_information(GIANT_BIT)[1] == pytest.approx(3.0)
     assert mi.s_information(XOR_DIST)[1] == pytest.approx(3.0)
-    
-    ptw, avg = mi.s_information(SUMMED_DICE)
-    
-    assert avg == pytest.approx(8.444326920731084)
-    assert sum(SUMMED_DICE[key]*ptw[key] for key in ptw.keys()) == pytest.approx(avg)
 
-    
+    ptw, avg = mi.s_information(SUMMED_DICE)
+
+    assert avg == pytest.approx(8.444326920731084)
+    assert sum(SUMMED_DICE[key] * ptw[key] for key in ptw.keys()) == pytest.approx(avg)
+
+
 def test_o_information():
-    
     assert mi.o_information(GIANT_BIT)[1] == pytest.approx(1.0)
     assert mi.o_information(XOR_DIST)[1] == pytest.approx(-1.0)
-    
+
     ptw, avg = mi.o_information(SUMMED_DICE)
-    
+
+    # From DIT
[…]
