+++ syntropy/lattices/utils.py
@@ -1,0 +1,24 @@
+import os
+import pathlib
+import pickle
+import numpy as np 
+import networkx as nx
+
+lattice_path = pathlib.Path(__file__).parent
+
+def load_lattice(num_inputs: int, num_target: int = 1):
+
+    if num_target == 1:
+        path = lattice_path / f"pi_lattice_{str(num_inputs)}.pickle"
+    elif num_target > 1:
+        path = lattice_path / f"pi_lattice_{str(num_inputs)+str(num_target)}.pickle"
+
+    with open(path, 'rb') as f:
+        lattice = pickle.load(f)
+    
+    return lattice
+         
+
+
+
+    
