+++ tests/test_gaussian.py
@@ -9,7 +9,6 @@ import syntropy.gaussian.shannon as shannon
 import syntropy.gaussian.multivariate_mi as mi
 from syntropy.gaussian.decompositions import (
     partial_information_decomposition as pid,
-    integrated_information_decomposition as phiid,
     partial_entropy_decomposition as ped,
     generalized_information_decomposition as gid,
     idep_partial_information_decomposition as idep,
@@ -17,7 +16,6 @@ from syntropy.gaussian.decompositions import (
 from syntropy.gaussian.temporal import (
     differential_entropy_rate,
     mutual_information_rate,
-    total_correlation_rate,
     o_information_rate,
 )
 
@@ -29,6 +27,49 @@ cov = np.cov(data, ddof=0.0)
 # Due to natural instability in Scipy's matrix algebra, we need a slightly
 # more relaxed tolerance for our unit tests. 1 part in 1,000,000 is probably ok.
 pytest_abs = 1e-6
+
+# Various helpers
+def equicorr_matrix(N: int, rho: float) -> np.ndarray:
+    """N x N equicorrelation matrix with off-diagonal rho."""
+    return (1 - rho) * np.eye(N) + rho * np.ones((N, N))
+
+
+def analytic_tc(N: int, rho: float) -> float:
+    """
+    TC(N, rho) = -(N-1)/2 * log(1-rho) - 1/2 * log(1 + (N-1)*rho)
+
+    Derived from TC = (1/2)*log(prod_i sigma_i^2 / |Sigma|)
+    with sigma_i^2 = 1 and |Sigma| = (1-rho)^{N-1} * (1+(N-1)*rho).
+    """
+    return -((N - 1) / 2) * np.log(1 - rho) - 0.5 * np.log(1 + (N - 1) * rho)
+
+
+def analytic_o_information(N: int, rho: float) -> float:
+    """
+    Omega(N, rho) = (N-2)/2 * log((1+(N-1)*rho) / (1-rho))
+                   - N/2 * log(1 + (N-2)*rho)
+
+    Derived from Omega = (2-N)*TC(N,rho) + N*TC(N-1,rho).
+
+    Sign convention: positive = redundancy dominated,
+                     negative = synergy dominated.
+    """
+    if N == 2:
+        return 0.0
+    term1 = ((N - 2) / 2) * np.log((1 + (N - 1) * rho) / (1 - rho))
+    term2 = (N / 2) * np.log(1 + (N - 2) * rho)
+    return term1 - term2
+
+
+def analytic_dtc(N: int, rho: float) -> float:
+    """DTC = TC - Omega."""
+    return analytic_tc(N, rho) - analytic_o_information(N, rho)
+
+
+def analytic_s_information(N: int, rho: float) -> float:
+    """S = TC + DTC."""
+    return analytic_tc(N, rho) + analytic_dtc(N, rho)
+
 
 def test_differential_entropy():
     h_all = shannon.local_differential_entropy(data).mean()
@@ -53,70 +94,122 @@ def test_differential_entropy():
 
 
 def test_mutual_information():
-    i = shannon.mutual_information((0,), (1,), cov)
-    r, _ = stats.pearsonr(data[0, :], data[1, :])
+    for m in range(10):
+        for n in range(m):
+            i = shannon.mutual_information(idxs_x=(m,), idxs_y=(n,), cov=cov)
+            r, _ = stats.pearsonr(data[m, :], data[n, :])
 
-    assert i == pytest.approx(-np.log(1 - (r**2)) / 2, abs=pytest_abs)
+            assert i == pytest.approx(-np.log(1 - (r**2)) / 2, abs=pytest_abs)
 
-    lmi = shannon.local_mutual_information((0,), (1,), data)
[…]
