+++ docs/build/html/_sources/api/syntropy.gaussian.rst.txt
@@ -1,26 +1,0 @@
-syntropy.gaussian package
-=========================
-
-.. automodule:: syntropy.gaussian
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Submodules
-----------
-
-syntropy.gaussian.optimization module
--------------------------------------
-
-.. automodule:: syntropy.gaussian.optimization
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-syntropy.gaussian.utils module
-------------------------------
-
-.. automodule:: syntropy.gaussian.utils
-   :members:
-   :show-inheritance:
-   :undoc-members:
