+++ docs/build/html/_sources/api/syntropy.discrete.rst.txt
@@ -1,42 +1,0 @@
-syntropy.discrete package
-=========================
-
-.. automodule:: syntropy.discrete
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Submodules
-----------
-
-syntropy.discrete.alpha\_synergy module
----------------------------------------
-
-.. automodule:: syntropy.discrete.alpha_synergy
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-syntropy.discrete.optimization module
--------------------------------------
-
-.. automodule:: syntropy.discrete.optimization
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-syntropy.discrete.temporal module
----------------------------------
-
-.. automodule:: syntropy.discrete.temporal
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-syntropy.discrete.utils module
-------------------------------
-
-.. automodule:: syntropy.discrete.utils
-   :members:
-   :show-inheritance:
-   :undoc-members:
