+++ docs/build/html/_sources/contributing.rst.txt
@@ -1,15 +1,0 @@
-Contributing
-============
-
-If you are interested in contributing to Syntropy, there are several fronts for development.
-The primary area is the code itself (features, bugfixes, efficiency), but we are also interested in tutorials, educational material, and example notebooks that can serve as a reference to new users. 
-
-To-do
------
-
-The space of information theory is *vast*, and while Syntropy aims to be comprehensive, there's a lot that has been untouched. If you are interested in implementing any of the following, feel free:
-
-* GPU acceleration for the neural estimators. 
-* Mixed estimators (e.g. mutual information between discrete and continuous random variables).
-* The `i_sx` and `i_min` redundancy functions for KNN-based PID. 
-* Bug fixes in any existing implementations. 
