+++ docs/build/html/_sources/api/syntropy.rst.txt
@@ -1,20 +1,0 @@
-syntropy package
-================
-
-.. automodule:: syntropy
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   syntropy.discrete
-   syntropy.gaussian
-   syntropy.knn
-   syntropy.lattices
-   syntropy.neural
-   syntropy.mixed
