+++ docs/build/html/_sources/api/syntropy.lattices.rst.txt
@@ -1,7 +1,0 @@
-syntropy.lattices package
-=========================
-
-.. automodule:: syntropy.lattices
-   :members:
-   :show-inheritance:
-   :undoc-members:
