### 2026-02-25 6815f3a
Remove biggest lattices
