### 2026-02-25 e4755e5
More tests + optimizations
