### 2026-02-25 32a630d
Optimizations
