### 2026-03-27 #9407 @crvernon
:wave:  @artemyk and @jkbren - are either or both of you able to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
