### 2026-03-27 #9407 @cbueth
Thanks for checking in, @crvernon. From my side there have been no news since the last status update. I did a second round in https://github.com/thosvarley/syntropy/issues/3#issuecomment-3801234549 and it seems there has been little activity on the main branch, but I hav enot heard back from @thosvarley.
