### 2026-08-18 #9407 @cbueth
Since my last issues were responded with https://github.com/thosvarley/syntropy/issues/3#issuecomment-4877099242, I did not check in, as I have overseen any notification and did not get back to verifying the response yet. I will need to find time.
