### 2026-08-19 #9407 @riteshbhirud
I am waiting for the issue [#3](https://github.com/thosvarley/syntropy/issues/3)  discussion to be resolved as I had some overlapping comments with that. Once that's done I can give final approval
