### 2026-09-05 #9407 @thosvarley
I'm not sure which aspects of #3 @riteshbhirud is thinking about - what parts did I miss?
